@@ -1,13 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
   <PropertyGroup>
-    <TargetFramework>net7.0-windows10.0.19041.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <AssemblyName>Shapes</AssemblyName>
     <RootNamespace>Shapes</RootNamespace>
+    <SupportedOSPlatformVersion>10.0.19041.0</SupportedOSPlatformVersion>
   </PropertyGroup>
-
   <ItemGroup>
     <ProjectReference Include="..\Coordinates\Coordinates.csproj" />
   </ItemGroup>
-
 </Project>
