@@ -1,9 +1,6 @@
 ﻿using Coordinates;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Shapes
 {
@@ -14,12 +11,11 @@ namespace Shapes
         public virtual double Calculate3DDistanceWithin(Track track, bool useGPSAltitude, bool isReentranceAllowed)
         {
             double distance = 0.0;
-            List<List<Coordinate>> pointsWithIn = new List<List<Coordinate>>();
-            pointsWithIn.Add(new List<Coordinate>());
+            List<List<Coordinate>> pointsWithIn = [[]];
             int count = 0;
             for (int index = 0; index < track.TrackPoints.Count; index++)
             {
-                if (IsWithin(track.TrackPoints[index],useGPSAltitude))
+                if (IsWithin(track.TrackPoints[index], useGPSAltitude))
                 {
                     count++;
                     pointsWithIn.Last().Add(track.TrackPoints[index]);
@@ -27,7 +23,7 @@ namespace Shapes
                 else
                 {
                     if (isReentranceAllowed)
-                        pointsWithIn.Add(new List<Coordinate>());
+                        pointsWithIn.Add([]);
                     else
                     {
                         if (count > 0)
@@ -37,12 +33,12 @@ namespace Shapes
                 }
             }
             if (!isReentranceAllowed)
-                distance = CoordinateHelpers.Calculate3DDistanceBetweenPoints(pointsWithIn[0],useGPSAltitude);
+                distance = CoordinateHelpers.Calculate3DDistanceBetweenPoints(pointsWithIn[0], useGPSAltitude);
             else
             {
                 for (int index = 0; index < pointsWithIn.Count; index++)
                 {
-                    distance += CoordinateHelpers.Calculate3DDistanceBetweenPoints(pointsWithIn[index],useGPSAltitude);
+                    distance += CoordinateHelpers.Calculate3DDistanceBetweenPoints(pointsWithIn[index], useGPSAltitude);
                 }
             }
 
