@@ -1,16 +1,13 @@
 ﻿using Competition;
 using Competition.Validation;
 using Coordinates;
-using LoggerComponent;
+using LoggingConnector;
+using Microsoft.Extensions.Logging;
 using OfficeOpenXml;
 using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
 using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -20,6 +17,11 @@ namespace BLC2021
     {
 
         #region Properties
+
+        private readonly ILogger<BLC2021TaskSheet2> Logger = LogConnector.LoggerFactory.CreateLogger<BLC2021TaskSheet2>();
+
+        private readonly PilotMapping PilotMapping = new();
+
         private bool BatchMode
         {
             get; set;
@@ -75,7 +77,6 @@ namespace BLC2021
         {
             BatchMode = batchMode;
             InitializeComponent();
-
             Text = ToString();
 
         }
@@ -93,7 +94,6 @@ namespace BLC2021
 
         private void BLC2021TaskSheet2_Load(object sender, EventArgs e)
         {
-            logListView1.StartLogging();
             if (BatchMode)
                 ConfigureForBatchMode();
             else
@@ -112,24 +112,22 @@ namespace BLC2021
             }
             if (showDialog)
             {
-                using (FolderBrowserDialog folderBrowserDialog = new FolderBrowserDialog())
+                using FolderBrowserDialog folderBrowserDialog = new();
+                folderBrowserDialog.RootFolder = Environment.SpecialFolder.MyComputer;
+                folderBrowserDialog.UseDescriptionForTitle = true;
+                folderBrowserDialog.Description = "Select an output directory";
+                if (folderBrowserDialog.ShowDialog() == DialogResult.OK)
                 {
-                    folderBrowserDialog.RootFolder = Environment.SpecialFolder.MyComputer;
-                    folderBrowserDialog.UseDescriptionForTitle = true;
-                    folderBrowserDialog.Description = "Select an output directory";
-                    if (folderBrowserDialog.ShowDialog() == DialogResult.OK)
-                    {
-                        OutputDirectory = new DirectoryInfo(folderBrowserDialog.SelectedPath);
-                        Properties.Settings.Default.DefaultOutputDirectory = folderBrowserDialog.SelectedPath;
-                        Properties.Settings.Default.Save();
-                        if (!btSelectIGCFiles.Enabled)
-                            btSelectIGCFiles.Enabled = true;
-                    }
-                    else
-                    {
-                        Log(LogSeverityType.Error, "No output directory has been defined, an output directory must be configured to continue");
-                        btSelectIGCFiles.Enabled = false;
-                    }
+                    OutputDirectory = new DirectoryInfo(folderBrowserDialog.SelectedPath);
+                    Properties.Settings.Default.DefaultOutputDirectory = folderBrowserDialog.SelectedPath;
+                    Properties.Settings.Default.Save();
+                    if (!btSelectIGCFiles.Enabled)
+                        btSelectIGCFiles.Enabled = true;
+                }
+                else
+                {
+                    Logger?.LogError("No output directory has been defined, an output directory must be configured to continue");
+                    btSelectIGCFiles.Enabled = false;
                 }
             }
             else
@@ -198,43 +196,39 @@ namespace BLC2021
         }
 
 
-        private void Log(LogSeverityType logSeverity, string logMessage)
-        {
-            Logger.Log(this, logSeverity, logMessage);
-        }
         private async void btSelectIGCFiles_Click(object sender, EventArgs e)
         {
             if (BatchMode)
             {
-                OpenFileDialog openFileDialog = new OpenFileDialog();
-                openFileDialog.Title = "Select .igc files";
-                openFileDialog.Filter = "igc files (*.igc)|*.igc";
-                openFileDialog.CheckPathExists = true;
+                OpenFileDialog openFileDialog = new()
+                {
+                    Title = "Select .igc files",
+                    Filter = "igc files (*.igc)|*.igc",
+                    CheckPathExists = true
+                };
                 openFileDialog.CheckPathExists = true;
                 openFileDialog.Multiselect = true;
                 UseWaitCursor = true;
 
-                HesitationWaltzTask task4_1 = new HesitationWaltzTask();
+                HesitationWaltzTask task4_1 = new();
                 task4_1.SetupHWZ(4, CalculateGoalsTask4_1, Task4_FirstMarkerNumber, true, null);
 
-                HesitationWaltzTask task4_2 = new HesitationWaltzTask();
+                HesitationWaltzTask task4_2 = new();
                 task4_2.SetupHWZ(4, CalculateGoalsTask4_2, Task4_SecondMarkerNumber, true, null);
 
-                HesitationWaltzTask task4_3 = new HesitationWaltzTask();
+                HesitationWaltzTask task4_3 = new();
                 task4_3.SetupHWZ(4, CalculateGoalsTask4_3, Task4_ThridMarkerNumber, true, null);
 
-                MarkerTimingRule markerTimingRule = new MarkerTimingRule();
-                markerTimingRule.SetupRule(new List<(int openAtMinute, int closeAtMinute)>() { (0, 5), (20, 25), (40, 45) });
+                MarkerTimingRule markerTimingRule = new();
+                markerTimingRule.SetupRule([(0, 5), (20, 25), (40, 45)]);
 
 
-                MarkerToGoalDistanceRule markerToGoalDistanceRule = new MarkerToGoalDistanceRule();
+                MarkerToGoalDistanceRule markerToGoalDistanceRule = new();
                 markerToGoalDistanceRule.SetupRule(double.NaN, 300.0, true, true, Task5_GoalNumber);
 
 
-                List<IMarkerValidationRules> markerValidationRules = new List<IMarkerValidationRules>();
-                markerValidationRules.Add(markerTimingRule);
-                markerValidationRules.Add(markerToGoalDistanceRule);
-                HesitationWaltzTask task5 = new HesitationWaltzTask();
+                List<IMarkerValidationRules> markerValidationRules = [markerTimingRule, markerToGoalDistanceRule];
+                HesitationWaltzTask task5 = new();
                 task5.SetupHWZ(5, CalculateGoalsTask5, Task5_MarkerNumber, true, markerValidationRules);
 
 
@@ -244,11 +238,11 @@ namespace BLC2021
                     bool success = await ParseFilesAndCalculateResultsAsync(igcFiles, task4_1, task4_2, task4_3, task5);
                     if (success)
                     {
-                        Log(LogSeverityType.Info, "Parsing and result calculation completed");
+                        Logger?.LogError("Parsing and result calculation completed");
                     }
                     else
                     {
-                        Log(LogSeverityType.Error, "Failed to parse and calculate results");
+                        Logger?.LogError("Failed to parse and calculate results");
                     }
                 }
                 UseWaitCursor = false;
@@ -265,7 +259,7 @@ namespace BLC2021
 
         private List<Coordinate> CalculateGoalsTask4_1(Track track)
         {
-            List<Coordinate> goals = new List<Coordinate>();
+            List<Coordinate> goals = [];
             Declaration declaration = track.GetLatestDeclaration(Task4_GoalNumber);
             if (declaration is null || declaration == default)
                 throw new Exception($"No goal with goal number '{Task4_GoalNumber}' has been found");
@@ -274,14 +268,14 @@ namespace BLC2021
         }
         private List<Coordinate> CalculateGoalsTask4_2(Track track)
         {
-            List<Coordinate> goals = new List<Coordinate>();
+            List<Coordinate> goals = [];
             double distanceFromDeclaredGoal = 1000.0;
             double additionalAltitude = CoordinateHelpers.ConvertToMeter(300.0);
             Declaration declaration = track.GetLatestDeclaration(Task4_GoalNumber);
             if (declaration is null || declaration == default)
                 throw new Exception($"No goal with goal number '{Task4_GoalNumber}' has been found");
 
-            CoordinateSharp.Coordinate tempDeclaredGoal = new CoordinateSharp.Coordinate(declaration.DeclaredGoal.Latitude, declaration.DeclaredGoal.Longitude);
+            CoordinateSharp.Coordinate tempDeclaredGoal = new(declaration.DeclaredGoal.Latitude, declaration.DeclaredGoal.Longitude);
 
             double easting = Math.Round(tempDeclaredGoal.UTM.Easting, 0, MidpointRounding.AwayFromZero);
             double northing = Math.Round(tempDeclaredGoal.UTM.Northing, 0, MidpointRounding.AwayFromZero);
@@ -294,7 +288,7 @@ namespace BLC2021
                 {
                     if (northingFactor == 0 && eastingFactor == 0)
                         continue;
-                    CoordinateSharp.UniversalTransverseMercator utmTempGoal = new CoordinateSharp.UniversalTransverseMercator($"{latZone}{longZone}", easting + (distanceFromDeclaredGoal * eastingFactor), northing + (distanceFromDeclaredGoal * northingFactor));
+                    CoordinateSharp.UniversalTransverseMercator utmTempGoal = new($"{latZone}{longZone}", easting + (distanceFromDeclaredGoal * eastingFactor), northing + (distanceFromDeclaredGoal * northingFactor));
                     CoordinateSharp.Coordinate tempGoal = CoordinateSharp.UniversalTransverseMercator.ConvertUTMtoLatLong(utmTempGoal);
                     goals.Add(new Coordinate(tempGoal.Latitude.DecimalDegree, tempGoal.Longitude.DecimalDegree, declaration.DeclaredGoal.AltitudeGPS + additionalAltitude, declaration.DeclaredGoal.AltitudeBarometric + additionalAltitude, DateTime.Now));
                 }
@@ -304,14 +298,14 @@ namespace BLC2021
 
         private List<Coordinate> CalculateGoalsTask4_3(Track track)
         {
-            List<Coordinate> goals = new List<Coordinate>();
+            List<Coordinate> goals = [];
             double distanceFromDeclaredGoal = 2000.0;
             double additionalAltitude = CoordinateHelpers.ConvertToMeter(600.0);
             Declaration declaration = track.GetLatestDeclaration(Task4_GoalNumber);
             if (declaration is null || declaration == default)
                 throw new Exception($"No goal with goal number '{Task4_GoalNumber}' has been found");
 
-            CoordinateSharp.Coordinate tempDeclaredGoal = new CoordinateSharp.Coordinate(declaration.DeclaredGoal.Latitude, declaration.DeclaredGoal.Longitude);
+            CoordinateSharp.Coordinate tempDeclaredGoal = new(declaration.DeclaredGoal.Latitude, declaration.DeclaredGoal.Longitude);
 
             double easting = Math.Round(tempDeclaredGoal.UTM.Easting, 0, MidpointRounding.AwayFromZero);
             double northing = Math.Round(tempDeclaredGoal.UTM.Northing, 0, MidpointRounding.AwayFromZero);
@@ -324,7 +318,7 @@ namespace BLC2021
                 {
                     if (northingFactor == 0 && eastingFactor == 0)
                         continue;
-                    CoordinateSharp.UniversalTransverseMercator utmTempGoal = new CoordinateSharp.UniversalTransverseMercator($"{latZone}{longZone}", easting + (distanceFromDeclaredGoal * eastingFactor), northing + (distanceFromDeclaredGoal * northingFactor));
+                    CoordinateSharp.UniversalTransverseMercator utmTempGoal = new($"{latZone}{longZone}", easting + (distanceFromDeclaredGoal * eastingFactor), northing + (distanceFromDeclaredGoal * northingFactor));
                     CoordinateSharp.Coordinate tempGoal = CoordinateSharp.UniversalTransverseMercator.ConvertUTMtoLatLong(utmTempGoal);
                     goals.Add(new Coordinate(tempGoal.Latitude.DecimalDegree, tempGoal.Longitude.DecimalDegree, declaration.DeclaredGoal.AltitudeGPS + additionalAltitude, declaration.DeclaredGoal.AltitudeBarometric + additionalAltitude, DateTime.Now));
                 }
@@ -335,7 +329,7 @@ namespace BLC2021
         private List<Coordinate> CalculateGoalsTask5(Track track)
         {
             {
-                List<Coordinate> goals = new List<Coordinate>();
+                List<Coordinate> goals = [];
                 Declaration declaration = track.GetLatestDeclaration(Task5_GoalNumber);
                 if (declaration is null || declaration == default)
                     throw new Exception($"No goal with goal number '{Task5_GoalNumber}' has been found");
@@ -348,48 +342,41 @@ namespace BLC2021
         {
             return await Task.Run(() =>
             {
-                string functionErrorMessage = "Failed to parse files and calculate results: ";
-                Track track;
-                List<(string pilotIdenifier, int pilotNumber, double result_T4, double result_T5)> results = new List<(string pilotIdenifier, int pilotNumber, double result_T4, double result_T5)>();
+                List<(string pilotIdenifier, int pilotNumber, double result_T4, double result_T5)> results = [];
                 try
                 {
-                    DeclarationToGoalDistanceRule declarationToGoalDistanceRule = new DeclarationToGoalDistanceRule();
+                    DeclarationToGoalDistanceRule declarationToGoalDistanceRule = new();
                     declarationToGoalDistanceRule.SetupRule(5000.0, double.NaN);
-                    DeclarationToGoalHeightRule declarationToGoalHeightRule = new DeclarationToGoalHeightRule();
+                    DeclarationToGoalHeightRule declarationToGoalHeightRule = new();
                     declarationToGoalHeightRule.SetupRule(CoordinateHelpers.ConvertToMeter(1000), double.NaN, DeclarationToGoalHeightRule.HeightDifferenceType.PositiveDifferenceOnly, true);
-                    List<IDeclarationValidationRules> declarationValidationRules = new List<IDeclarationValidationRules>();
-                    declarationValidationRules.Add(declarationToGoalDistanceRule);
-                    declarationValidationRules.Add(declarationToGoalHeightRule);
+                    List<IDeclarationValidationRules> declarationValidationRules = [declarationToGoalDistanceRule, declarationToGoalHeightRule];
                     foreach (string igcFile in igcFiles)
                     {
-                        if (!Coordinates.Parsers.BalloonLiveParser.ParseFile(igcFile, out track))
+                        if (!Coordinates.Parsers.BalloonLiveParser.ParseFile(igcFile, out Track track))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse '{igcFile}'");
+                            Logger?.LogError("Failed to parse files and calculate results: Failed to parse '{igcFile}'", igcFile);
                             continue;
                         }
                         double result_Task4;
                         if (ValidationHelper.GetValidDeclaration(track, Task4_GoalNumber, declarationValidationRules) != null)
                         {
-                            double result_Task4_1;
-                            if (!task4_1.CalculateResults(track, true, out result_Task4_1))
+                            if (!task4_1.CalculateResults(track, true, out double result_Task4_1))
                             {
-                                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to calculate result at Task 4 for Pilot No {track.Pilot.PilotNumber}");
+                                Logger?.LogError("Failed to parse files and calculate results: Failed to calculate result at Task 4 for Pilot No '{pilotNumber}'", track.Pilot.PilotNumber);
                                 result_Task4_1 = double.NaN;
                             }
                             task4_1.Goals.Clear();
 
-                            double result_Task4_2;
-                            if (!task4_2.CalculateResults(track, true, out result_Task4_2))
+                            if (!task4_2.CalculateResults(track, true, out double result_Task4_2))
                             {
-                                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to calculate result at Task 4 for Pilot No {track.Pilot.PilotNumber}");
+                                Logger?.LogError("Failed to parse files and calculate results: Failed to calculate result at Task 4 for Pilot No '{pilotNumber}'", track.Pilot.PilotNumber);
                                 result_Task4_2 = double.NaN;
                             }
                             task4_2.Goals.Clear();
 
-                            double result_Task4_3;
-                            if (!task4_3.CalculateResults(track, true, out result_Task4_3))
+                            if (!task4_3.CalculateResults(track, true, out double result_Task4_3))
                             {
-                                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to calculate result at Task 4 for Pilot No {track.Pilot.PilotNumber}");
+                                Logger?.LogError("Failed to parse files and calculate results: Failed to calculate result at Task 4 for Pilot No '{pilotNumber}'", track.Pilot.PilotNumber);
                                 result_Task4_3 = double.NaN;
                             }
                             task4_3.Goals.Clear();
@@ -397,7 +384,7 @@ namespace BLC2021
                             if (!double.IsNaN(result_Task4_1) && !double.IsNaN(result_Task4_2) && !double.IsNaN(result_Task4_3))
                             {
                                 result_Task4 = result_Task4_1 + result_Task4_2 + result_Task4_3;
-                                Log(LogSeverityType.Info, $"Calculated result of '{Math.Round(result_Task4, 0, MidpointRounding.AwayFromZero)}m' ({result_Task4_1:0.#}+{result_Task4_2:0.#}+{result_Task4_3:0.#}) at Task 4 for Pilot No {track.Pilot.PilotNumber}");
+                                Logger?.LogInformation("Calculated result of '{result_Task4}m' ({result_Task4_1} + {result_Task4_2} + {result_Task4_3}) at Task 4 for Pilot No '{pilotNumber}'", Math.Round(result_Task4, 0, MidpointRounding.AwayFromZero), $"{result_Task4_1: 0.#}", $"{result_Task4_2: 0.#}", $"{result_Task4_3:0.#}", track.Pilot.PilotNumber);
                             }
                             else
                             {
@@ -416,12 +403,12 @@ namespace BLC2021
                             task5.MarkerValidationRules.OfType<MarkerToGoalDistanceRule>().First().Declaration = declaration;
                             if (!task5.CalculateResults(track, true, out result_Task5))
                             {
-                                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to calculate result at Task 5 for Pilot No {track.Pilot.PilotNumber}");
+                                Logger?.LogError("Failed to parse files and calculate results: Failed to calculate result at Task 5 for Pilot No '{pilotNumber}'", track.Pilot.PilotNumber);
                                 result_Task5 = double.NaN;
                             }
                             else
                             {
-                                Log(LogSeverityType.Info, $"Calculated result of '{Math.Round(result_Task5, 3, MidpointRounding.AwayFromZero)}m' at Task 5 for Pilot No {track.Pilot.PilotNumber}");
+                                Logger?.LogInformation("Calculated result of '{result_Task5}m' at Task 5 for Pilot No '{pilotNumber}'", Math.Round(result_Task5, 0, MidpointRounding.AwayFromZero), track.Pilot.PilotNumber);
                             }
                             task5.Goals.Clear();
                         }
@@ -432,9 +419,9 @@ namespace BLC2021
                         results.Add((track.Pilot.PilotIdentifier, track.Pilot.PilotNumber, result_Task4, result_Task5));
                     }
 
-                    ExcelPackage.LicenseContext = OfficeOpenXml.LicenseContext.NonCommercial;
-                    FileInfo resultsFileInternal = new FileInfo(Path.Combine(OutputDirectory.FullName, ResultsFileNameInternal));
-                    using (ExcelPackage package = new ExcelPackage(resultsFileInternal))
+                    ExcelPackage.LicenseContext = LicenseContext.NonCommercial;
+                    FileInfo resultsFileInternal = new(Path.Combine(OutputDirectory.FullName, ResultsFileNameInternal));
+                    using (ExcelPackage package = new(resultsFileInternal))
                     {
                         ExcelWorksheet wsResults = package.Workbook.Worksheets.FirstOrDefault(x => x.Name == "Results");
                         if (wsResults == default)
@@ -446,24 +433,24 @@ namespace BLC2021
                             wsResults.Cells[1, 4].Value = "Task 5 CRT [m]";
                             wsResults.Cells[1, 5].Value = "Task 4 Comment";
                             wsResults.Cells[1, 6].Value = "Task 5 Comment";
-                            
+
                         }
-                        foreach ((string pilotIdentifier, int pilotNumber, double result_T4, double result_T5) result in results)
+                        foreach ((string pilotIdentifier, int pilotNumber, double result_T4, double result_T5) in results)
                         {
-                            wsResults.Cells[result.pilotNumber + 1, 1].Value = result.pilotIdentifier;
-                            wsResults.Cells[result.pilotNumber + 1, 2].Value = result.pilotNumber;
-                            wsResults.Cells[result.pilotNumber + 1, 3].Value = Math.Round(result.result_T4, 0, MidpointRounding.AwayFromZero);
-                            wsResults.Cells[result.pilotNumber + 1, 4].Value = Math.Round(result.result_T5 , 0, MidpointRounding.AwayFromZero);
-                            
-                            wsResults.Cells[result.pilotNumber + 1, 3].Style.Fill.PatternType = OfficeOpenXml.Style.ExcelFillStyle.None;
-                            wsResults.Cells[result.pilotNumber + 1, 4].Style.Fill.PatternType = OfficeOpenXml.Style.ExcelFillStyle.None;
+                            wsResults.Cells[pilotNumber + 1, 1].Value = pilotIdentifier;
+                            wsResults.Cells[pilotNumber + 1, 2].Value = pilotNumber;
+                            wsResults.Cells[pilotNumber + 1, 3].Value = Math.Round(result_T4, 0, MidpointRounding.AwayFromZero);
+                            wsResults.Cells[pilotNumber + 1, 4].Value = Math.Round(result_T5, 0, MidpointRounding.AwayFromZero);
+
+                            wsResults.Cells[pilotNumber + 1, 3].Style.Fill.PatternType = OfficeOpenXml.Style.ExcelFillStyle.None;
+                            wsResults.Cells[pilotNumber + 1, 4].Style.Fill.PatternType = OfficeOpenXml.Style.ExcelFillStyle.None;
                         }
                         wsResults.Cells.AutoFitColumns();
                         package.Save();
                     }
-                    Log(LogSeverityType.Info, $"Successfully created or modified internal results file '{resultsFileInternal.Name}'");
-                    FileInfo resultsFileProvisional = new FileInfo(Path.Combine(OutputDirectory.FullName, ResultsFileNameProvisional));
-                    using (ExcelPackage package = new ExcelPackage(resultsFileProvisional))
+                    Logger?.LogInformation("Successfully created or modified internal results file '{resultsFileInternal.Name}'", resultsFileInternal.Name);
+                    FileInfo resultsFileProvisional = new(Path.Combine(OutputDirectory.FullName, ResultsFileNameProvisional));
+                    using (ExcelPackage package = new(resultsFileProvisional))
                     {
                         ExcelWorksheet wsResults = package.Workbook.Worksheets.FirstOrDefault(x => x.Name == "Results");
                         if (wsResults == default)
@@ -475,38 +462,36 @@ namespace BLC2021
                             wsResults.Cells[1, 4].Value = "Pilot First Name";
                             wsResults.Cells[1, 5].Value = "Task 4 PDG [m]";
                             wsResults.Cells[1, 6].Value = "Task 5 CRT [m]";
-                            
+
                         }
-                        string firstName;
-                        string lastName;
-                        foreach ((string pilotIdentifier, int pilotNumber, double result_T4, double result_T5) result in results)
+                        foreach ((string pilotIdentifier, int pilotNumber, double result_T4, double result_T5) in results)
                         {
-                            wsResults.Cells[result.pilotNumber + 1, 1].Value = result.pilotIdentifier;
-                            wsResults.Cells[result.pilotNumber + 1, 2].Value = result.pilotNumber;
-                            if (!PilotMapping.Instance.GetPilotName(result.pilotNumber, out lastName, out firstName))
+                            wsResults.Cells[pilotNumber + 1, 1].Value = pilotIdentifier;
+                            wsResults.Cells[pilotNumber + 1, 2].Value = pilotNumber;
+                            if (!PilotMapping.GetPilotName(pilotNumber, out string lastName, out string firstName))
                             {
-                                Log(LogSeverityType.Warning, "Last name and first name of pilot will be omitted");
+                                Logger?.LogWarning("Last name and first name of pilot will be omitted");
                             }
                             else
                             {
-                                wsResults.Cells[result.pilotNumber + 1, 3].Value = lastName;
-                                wsResults.Cells[result.pilotNumber + 1, 4].Value = firstName;
+                                wsResults.Cells[pilotNumber + 1, 3].Value = lastName;
+                                wsResults.Cells[pilotNumber + 1, 4].Value = firstName;
                             }
-                            wsResults.Cells[result.pilotNumber + 1, 5].Value = Math.Round(result.result_T4, 0, MidpointRounding.AwayFromZero);
-                            wsResults.Cells[result.pilotNumber + 1, 6].Value = Math.Round(result.result_T5 , 0, MidpointRounding.AwayFromZero);
-                            
+                            wsResults.Cells[pilotNumber + 1, 5].Value = Math.Round(result_T4, 0, MidpointRounding.AwayFromZero);
+                            wsResults.Cells[pilotNumber + 1, 6].Value = Math.Round(result_T5, 0, MidpointRounding.AwayFromZero);
 
-                            wsResults.Cells[result.pilotNumber + 1, 5].Style.Fill.PatternType = OfficeOpenXml.Style.ExcelFillStyle.None;
-                            wsResults.Cells[result.pilotNumber + 1, 6].Style.Fill.PatternType = OfficeOpenXml.Style.ExcelFillStyle.None;
+
+                            wsResults.Cells[pilotNumber + 1, 5].Style.Fill.PatternType = OfficeOpenXml.Style.ExcelFillStyle.None;
+                            wsResults.Cells[pilotNumber + 1, 6].Style.Fill.PatternType = OfficeOpenXml.Style.ExcelFillStyle.None;
                         }
                         wsResults.Cells.AutoFitColumns();
                         package.Save();
                     }
-                    Log(LogSeverityType.Info, $"Successfully created or modified provisional results file '{resultsFileProvisional.Name}'");
+                    Logger?.LogInformation("Successfully created or modified provisional results file '{resultsFileProvisional.Name}'", resultsFileProvisional.Name);
                 }
                 catch (Exception ex)
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"{ex.Message}");
+                    Logger?.LogError(ex, "Failed to parse files and calculate results");
                     return false;
                 }
 
