@@ -1,38 +1,23 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.IO;
+﻿using LoggingConnector;
+using Microsoft.Extensions.Logging;
 using OfficeOpenXml;
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 using System.Windows.Forms;
-using LoggerComponent;
-using System.Threading;
 
 namespace BLC2021
 {
     public sealed class PilotMapping
     {
-        private static readonly Lazy<PilotMapping> lazy =
-                new Lazy<PilotMapping>(() => new PilotMapping());
-
-        public static PilotMapping Instance
-        {
-            get
-            {
-                return lazy.Value;
-            }
-        }
-
+        private readonly ILogger<PilotMapping> Logger = LogConnector.LoggerFactory.CreateLogger<PilotMapping>();
 
         private List<(int pilotNumber, string lastName, string firstName)> PilotMappings
         {
             get; set;
         }
 
-        private PilotMapping()
-        {
-        }
 
         public bool GetPilotName(int pilotNumber, out string lastName, out string firstName)
         {
@@ -42,7 +27,7 @@ namespace BLC2021
             {
                 if (!LoadPilotMapping())
                 {
-                    Log(LogSeverityType.Error, $"Failed to get pilot name for pilot number '{pilotNumber}'");
+                    Logger?.LogError("Failed to get pilot name for pilot number '{pilotNumber}'", pilotNumber);
                     return false;
                 }
             }
@@ -50,7 +35,7 @@ namespace BLC2021
 
             if (pilotMapping == default)
             {
-                Log(LogSeverityType.Warning, $"Failed to get pilot name for pilot number '{pilotNumber}' : Pilot number not found within mappings");
+                Logger?.LogError("Failed to get pilot name for pilot number '{pilotNumber}' : Pilot number not found within mappings", pilotNumber);
                 return false;
 
             }
@@ -75,32 +60,28 @@ namespace BLC2021
             if (showDialog)
             {
 
-                Thread t = new Thread((ThreadStart)(() => {
-                    OpenFileDialog openFileDialog = new OpenFileDialog();
-                openFileDialog.AddExtension = true;
-                openFileDialog.CheckFileExists = true;
-                openFileDialog.CheckPathExists = true;
-                openFileDialog.Multiselect = false;
-                openFileDialog.Title = "Select pilot mapping";
-                openFileDialog.Filter = "xlsx files (*.xlsx)|*.xlsx";
+                OpenFileDialog openFileDialog = new()
+                {
+                    AddExtension = true,
+                    CheckFileExists = true,
+                    CheckPathExists = true,
+                    Multiselect = false,
+                    Title = "Select pilot mapping",
+                    Filter = "xlsx files (*.xlsx)|*.xlsx"
+                };
                 if (openFileDialog.ShowDialog() == DialogResult.OK)
                 {
                     pilotMappingFile = new FileInfo(openFileDialog.FileName);
                     Properties.Settings.Default.PathToPilotMapping = pilotMappingFile.FullName;
                     Properties.Settings.Default.Save();
                 }
-                }));
 
-                
-                t.SetApartmentState(ApartmentState.STA);
-                t.Start();
-                t.Join();
             }
             try
             {
-                List<(int pilotNumber, string lastName, string firstName)> pilotMappings = new List<(int pilotNumber, string lastName, string firstName)>();
+                List<(int pilotNumber, string lastName, string firstName)> pilotMappings = [];
                 ExcelPackage.LicenseContext = OfficeOpenXml.LicenseContext.NonCommercial;
-                using (ExcelPackage package = new ExcelPackage(pilotMappingFile))
+                using (ExcelPackage package = new(pilotMappingFile))
                 {
                     ExcelWorksheet wsPilots = package.Workbook.Worksheets.First();
 
@@ -124,12 +105,11 @@ namespace BLC2021
                     }
                 }
                 PilotMappings = pilotMappings;
-                Log(LogSeverityType.Info, "Pilot mappings successfully loaded");
-
+                Logger?.LogInformation("Pilot mappings successfully loaded");
             }
             catch (Exception)
             {
-                Log(LogSeverityType.Error, "Failed to load pilot mapping");
+                Logger?.LogError("Failed to load pilot mapping");
                 return false;
             }
             return true;
@@ -139,11 +119,6 @@ namespace BLC2021
         public override string ToString()
         {
             return "PilotMapping";
-        }
-
-        private void Log(LogSeverityType logSeverity, string logMessage)
-        {
-            Logger.Log(this, logSeverity, logMessage);
         }
     }
 }
