@@ -1,21 +1,22 @@
 ﻿using BalloonTrackAnalyze.TaskControls;
 using Competition;
 using Coordinates;
-using LoggerComponent;
+using LoggingConnector;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
 using System.Drawing;
 using System.IO;
-using System.Text;
 using System.Windows.Forms;
 
 namespace BalloonTrackAnalyze
 {
     public partial class Form1 : Form
     {
+        private readonly ILogger<Form1> Logger = LogConnector.LoggerFactory.CreateLogger<Form1>();
+
+
         private Point TaskControlLocation { get; set; } = new Point(0, 0);
 
         private Flight flight
@@ -32,23 +33,22 @@ namespace BalloonTrackAnalyze
 
         private void Form1_Load(object sender, EventArgs e)
         {
-            logListView1.StartLogging(@".\logfile.txt", 5);
             if (!string.IsNullOrWhiteSpace(Properties.Settings.Default.CompetitionFolder))
             {
                 tbCompetitionFolder.Text = Properties.Settings.Default.CompetitionFolder;
-                Log(LogSeverityType.Info, $"Loaded Competition Folder '{tbCompetitionFolder.Text}' from settings");
+                Logger?.LogInformation("Loaded Competition Folder '{competitionFolder}' from settings", tbCompetitionFolder.Text);
             }
             if (!string.IsNullOrWhiteSpace(Properties.Settings.Default.PilotNameMappingFile))
             {
                 tbPilotMappingFile.Text = Properties.Settings.Default.PilotNameMappingFile;
-                Log(LogSeverityType.Info, $"Loaded Pilot Name Mapping File '{tbPilotMappingFile.Text}' from settings");
+                Logger?.LogInformation("Loaded Pilot Name Mapping File '{pilotNameMappingFile}' from settings", tbPilotMappingFile.Text);
             }
             rbBalloonLive.Checked = Properties.Settings.Default.UseBalloonLiveParser;
             rbFAILogger.Checked = !Properties.Settings.Default.UseBalloonLiveParser;
-            Log(LogSeverityType.Info, $"Loaded Track Source '{(Properties.Settings.Default.UseBalloonLiveParser ? "Balloon Live" : "FAI Logger")}' from settings");
+            Logger?.LogInformation("Loaded Track Source '{trackSource}' from settings", (Properties.Settings.Default.UseBalloonLiveParser ? "Balloon Live" : "FAI Logger"));
             rbGPSAltitude.Checked = Properties.Settings.Default.UseGPSAltitude;
             rbBarometricAltitude.Checked = !Properties.Settings.Default.UseGPSAltitude;
-            Log(LogSeverityType.Info, $"Loaded Height Source '{(Properties.Settings.Default.UseGPSAltitude ? "GPS" : "Barometric")}' from settings");
+            Logger?.LogInformation("Loaded Height Source '{heightSource}' from settings", (Properties.Settings.Default.UseGPSAltitude ? "GPS" : "Barometric"));
         }
 
         private void cbTaskList_SelectedIndexChanged(object sender, EventArgs e)
@@ -69,7 +69,7 @@ namespace BalloonTrackAnalyze
                     break;
                 case "Pie":
                     {
-                        PieControl pieControl = new PieControl();
+                        PieControl pieControl = new();
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         pieControl.Location = TaskControlLocation;
@@ -81,7 +81,7 @@ namespace BalloonTrackAnalyze
                     break;
                 case "Elbow":
                     {
-                        ElbowControl elbowControl = new ElbowControl();
+                        ElbowControl elbowControl = new();
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         elbowControl.Location = TaskControlLocation;
@@ -93,7 +93,7 @@ namespace BalloonTrackAnalyze
                     break;
                 case "Landrun":
                     {
-                        LandRunControl landrunControl = new LandRunControl();
+                        LandRunControl landrunControl = new();
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         landrunControl.Location = TaskControlLocation;
@@ -111,11 +111,6 @@ namespace BalloonTrackAnalyze
             }
         }
 
-        private void Log(LogSeverityType logSeverityType, string text)
-        {
-            Logger.Log(this, logSeverityType, text);
-        }
-
         private void LandrunControl_DataValid()
         {
             if (plUserControl.Controls["taskControl"] is LandRunControl landRunControl)
@@ -125,20 +120,20 @@ namespace BalloonTrackAnalyze
                 {
                     if (DoesTaskNumberAlreadyExists(landRun.TaskNumber))
                     {
-                        Log(LogSeverityType.Error, $"Failed to create {landRun}: A task with the number '{landRun.TaskNumber}' already exists");
+                        Logger?.LogError("Failed to create '{landRun}': A task with the number '{landRun.TaskNumber}' already exists", landRun, landRun.TaskNumber);
                     }
                     else
                     {
                         if (!lbTaskList.Items.Contains(landRun))
                         {
                             lbTaskList.Items.Add(landRun);
-                            Log(LogSeverityType.Info, $"{landRun} created successfully");
+                            Logger?.LogInformation("{landRun} created successfully", landRun);
                         }
                     }
                 }
                 else
                 {
-                    Log(LogSeverityType.Info, $"{landRun} modified successfully");
+                    Logger?.LogInformation("{landRun} modified successfully", landRun);
                 }
             }
         }
@@ -152,20 +147,20 @@ namespace BalloonTrackAnalyze
                 {
                     if (DoesTaskNumberAlreadyExists(elbow.TaskNumber))
                     {
-                        Log(LogSeverityType.Error, $"Failed to create {elbow}: A task with the number '{elbow.TaskNumber}' already exists");
+                        Logger?.LogError("Failed to create {elbow}: A task with the number '{elbow.TaskNumber}' already exists", elbow, elbow.TaskNumber);
                     }
                     else
                     {
                         if (!lbTaskList.Items.Contains(elbow))
                         {
                             lbTaskList.Items.Add(elbow);
-                            Log(LogSeverityType.Info, $"{elbow} created successfully");
+                            Logger?.LogInformation("{elbow} created successfully", elbow);
                         }
                     }
                 }
                 else
                 {
-                    Log(LogSeverityType.Info, $"{elbow} modified successfully");
+                    Logger?.LogInformation("{elbow} modified successfully", elbow);
                 }
             }
         }
@@ -179,20 +174,20 @@ namespace BalloonTrackAnalyze
                 {
                     if (DoesTaskNumberAlreadyExists(pie.TaskNumber))
                     {
-                        Log(LogSeverityType.Error, $"Failed to create {pie}: A task with the number '{pie.TaskNumber}' already exists");
+                        Logger?.LogError("Failed to create {pie}: A task with the number '{pie.TaskNumber}' already exists", pie, pie.TaskNumber);
                     }
                     else
                     {
                         if (!lbTaskList.Items.Contains(pie))
                         {
                             lbTaskList.Items.Add(pie);
-                            Log(LogSeverityType.Info, $"{pie} created successfully");
+                            Logger?.LogInformation("{pie} created successfully", pie);
                         }
                     }
                 }
                 else
                 {
-                    Log(LogSeverityType.Info, $"{pie} modified successfully");
+                    Logger?.LogInformation("{pie} modified successfully", pie);
                 }
             }
         }
@@ -207,19 +202,21 @@ namespace BalloonTrackAnalyze
                 {
                     if (DoesTaskNumberAlreadyExists(donut.TaskNumber))
                     {
-                        Log(LogSeverityType.Error, $"Failed to create {donut}: A task with the number '{donut.TaskNumber}' already exists");
+                        Logger?.LogError("Failed to create {donut}: A task with the number '{donut.TaskNumber}' already exists", donut, donut.TaskNumber);
                     }
                     else
                     {
                         if (!lbTaskList.Items.Contains(donut))
                         {
                             lbTaskList.Items.Add(donut);
-                            Log(LogSeverityType.Info, $"{donut} created successfully");
+                            Logger?.LogInformation("{donut} created successfully", donut);
                         }
                     }
                 }
                 else
-                    Log(LogSeverityType.Info, $"{donut} modified successfully");
+                {
+                    Logger?.LogInformation("{donut} modified successfully", donut);
+                }
             }
         }
 
@@ -247,7 +244,7 @@ namespace BalloonTrackAnalyze
             Properties.Settings.Default.UseBalloonLiveParser = rbBalloonLive.Checked;
             Properties.Settings.Default.UseGPSAltitude = rbGPSAltitude.Checked;
             Properties.Settings.Default.Save();
-            Log(LogSeverityType.Info, "Saved competition settings");
+            Logger?.LogInformation("Saved competition settings");
         }
 
         private void btDeleteTask_Click(object sender, EventArgs e)
@@ -277,7 +274,7 @@ namespace BalloonTrackAnalyze
             {
                 case DonutTask donut:
                     {
-                        DonutControl donutControl = new DonutControl(donut);
+                        DonutControl donutControl = new(donut);
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         donutControl.Location = TaskControlLocation;
@@ -289,7 +286,7 @@ namespace BalloonTrackAnalyze
                     break;
                 case PieTask pie:
                     {
-                        PieControl pieControl = new PieControl(pie);
+                        PieControl pieControl = new(pie);
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         pieControl.Location = TaskControlLocation;
@@ -301,7 +298,7 @@ namespace BalloonTrackAnalyze
                     break;
                 case ElbowTask elbow:
                     {
-                        ElbowControl elbowControl = new ElbowControl(elbow);
+                        ElbowControl elbowControl = new(elbow);
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         elbowControl.Location = TaskControlLocation;
@@ -313,7 +310,7 @@ namespace BalloonTrackAnalyze
                     break;
                 case LandRunTask landRun:
                     {
-                        LandRunControl landrunControl = new LandRunControl(landRun);
+                        LandRunControl landrunControl = new(landRun);
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         landrunControl.Location = TaskControlLocation;
@@ -328,16 +325,14 @@ namespace BalloonTrackAnalyze
 
         private void btCalculateResults_Click(object sender, EventArgs e)
         {
-            string functionErrorMessage = "Failed to calculate Results :";
             if (string.IsNullOrWhiteSpace(tbFlightNumber.Text))
             {
-                Log(LogSeverityType.Error, functionErrorMessage + $"Please enter a Flight No first");
+                Logger?.LogError("Failed to calculate Results: Please enter a Flight No first");
                 return;
             }
-            int flightNumber;
-            if (!int.TryParse(tbFlightNumber.Text, out flightNumber))
+            if (!int.TryParse(tbFlightNumber.Text, out int flightNumber))
             {
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse Flight No '{tbFlightNumber.Text}' as integer");
+                Logger?.LogError("Failed to calculate Results: Failed to parse Flight No '{flightNumber}' as integer", tbFlightNumber.Text);
                 return;
             }
             CreateFolderStructure(flightNumber);
@@ -361,7 +356,7 @@ namespace BalloonTrackAnalyze
                     }
                     catch (Exception ex)
                     {
-                        Log(LogSeverityType.Warning, $"Failed to serialize task '{item}': {ex.Message} ");
+                        Logger?.LogWarning("Failed to serialize task '{item}': {ex.Message} ", item, ex.Message);
                     }
                 }
             }
@@ -369,53 +364,53 @@ namespace BalloonTrackAnalyze
             {
                 if (!flight.ParseTrackFiles(Path.Combine(tbCompetitionFolder.Text, "Scoring", $"Flight{flightNumber:D2}", "Tracks"), rbBalloonLive.Checked))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage);
+                    Logger?.LogError("Failed to calculate Results");
                     return;
                 }
-                Log(LogSeverityType.Info, $"{flight.Tracks.Count} Track files parsed successfully");
+                Logger?.LogInformation("{flight.Tracks.Count} Track files parsed successfully", flight.Tracks.Count);
             }
             catch (Exception ex)
             {
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse track files: {ex.Message}");
+                Logger?.LogError(ex, "Failed to calculate Results");
                 return;
             }
             try
             {
                 if (!flight.MapPilotNamesToTracks(tbPilotMappingFile.Text))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage);
+                    Logger?.LogError("Failed to calculate Results");
                     return;
                 }
-                Log(LogSeverityType.Info, "Successfully mapped pilot names to tracks");
+                Logger?.LogInformation("Successfully mapped pilot names to tracks");
             }
             catch (Exception ex)
             {
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to map pilot names to tracks: {ex.Message}");
+                Logger?.LogError(ex, "Failed to calculate Results: failed to map pilot names to tracks");
                 return;
             }
             try
             {
                 CreateMarkerAndDeclarationFiles(flightNumber);
-                Log(LogSeverityType.Info, "Successfully created file with markers and declarations for each pilot");
+                Logger?.LogInformation("Successfully created file with markers and declarations for each pilot");
             }
             catch (Exception ex)
             {
-                Log(LogSeverityType.Warning, functionErrorMessage + $"Failed to create marker and declaration files: {ex.Message}");
+                Logger?.LogError(ex, "Failed to calculate Results: failed to create marker and declaration files");
             }
             try
             {
                 flight.CalculateResults(rbGPSAltitude.Checked, Path.Combine(tbCompetitionFolder.Text, "Scoring", $"Flight{flightNumber:D2}"));
-                Log(LogSeverityType.Info, "Successfully calculated results of all pilots for each task");
+                Logger?.LogInformation("Successfully calculated results of all pilots for each task");
             }
             catch (Exception ex)
             {
-                Log(LogSeverityType.Warning, functionErrorMessage + $"Failed to calculate results: {ex.Message}");
+                Logger?.LogError(ex, "Failed to calculate Results: failed to calculate results of all pilots for each task");
             }
         }
 
         private void CreateFolderStructure(int flightNumber)
         {
-            DirectoryInfo directoryInfo = new DirectoryInfo($@"{tbCompetitionFolder.Text}");
+            DirectoryInfo directoryInfo = new($@"{tbCompetitionFolder.Text}");
             if (!directoryInfo.Exists)
                 directoryInfo.Create();
             directoryInfo = new DirectoryInfo($@"{Path.Combine(tbCompetitionFolder.Text, "Scoring")}");
@@ -439,48 +434,48 @@ namespace BalloonTrackAnalyze
         {
             foreach (Track track in flight.Tracks)
             {
-                using (StreamWriter writer = new StreamWriter($@"{Path.Combine(tbCompetitionFolder.Text, "Scoring", $"Flight{flightNumber:D2}", "Pilots", $"Pilot{track.Pilot.PilotNumber:D3}{(!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"_{track.Pilot.FirstName}_{track.Pilot.LastName}" : "")}.csv")}", false))
+                using StreamWriter writer = new($@"{Path.Combine(tbCompetitionFolder.Text, "Scoring", $"Flight{flightNumber:D2}", "Pilots", $"Pilot{track.Pilot.PilotNumber:D3}{(!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"_{track.Pilot.FirstName}_{track.Pilot.LastName}" : "")}.csv")}", false);
+                string markerHeader = "Marker Number, Time Stamp, Latitude[dec], Longitude[dec], Latitude[deg], Longitude[deg], UTM Zone, Easting, Norting, Altitude GPS[m], Altitude Barometric[m]";
+                writer.WriteLine(markerHeader);
+                foreach (MarkerDrop marker in track.MarkerDrops)
                 {
-                    string markerHeader = "Marker Number, Time Stamp, Latitude[dec], Longitude[dec], Latitude[deg], Longitude[deg], UTM Zone, Easting, Norting, Altitude GPS[m], Altitude Barometric[m]";
-                    writer.WriteLine(markerHeader);
-                    foreach (MarkerDrop marker in track.MarkerDrops)
-                    {
-                        string timeStamp = marker.MarkerLocation.TimeStamp.ToString("dd.MM.yyyy HH:mm:ss");
-                        CoordinateSharp.Coordinate coordinateSharp = new CoordinateSharp.Coordinate(marker.MarkerLocation.Latitude, marker.MarkerLocation.Longitude);
-                        string line = string.Join(',', marker.MarkerNumber, timeStamp, marker.MarkerLocation.Latitude, marker.MarkerLocation.Longitude,
-                            coordinateSharp.Latitude.Display, coordinateSharp.Longitude.Display, coordinateSharp.UTM.LongZone + coordinateSharp.UTM.LatZone,
-                            coordinateSharp.UTM.Easting, coordinateSharp.UTM.Northing, marker.MarkerLocation.AltitudeGPS, marker.MarkerLocation.AltitudeBarometric);
-                        writer.WriteLine(line);
-                    }
-                    writer.WriteLine();
-                    writer.WriteLine();
-                    writer.WriteLine();
-                    string declarationHeader = "Declaration Number,Time Stamp,Position Latitude[dec],Position Longitude[dec],Position Latitude[deg],Position Longitude[deg],Position Altitude GPS[m],Position Altitude Barometric[m],Position UTM Zone,Position Easting,Position Norting,Goal Latitude[dec],Goal Longitude[dec],Goal Latitude[deg],Goal Longitude[deg],Goal UTM Zone,Goal Easting,Goal Norting,Goal Altitude GPS m],Goal Altitude Barometric[m]";
-                    writer.WriteLine(declarationHeader);
-                    foreach (Declaration declaration in track.Declarations)
-                    {
-                        string timeStamp = declaration.PositionAtDeclaration.TimeStamp.ToString("dd.MM.yyyy HH:mm:ss");
-                        CoordinateSharp.Coordinate posCoordinate = new CoordinateSharp.Coordinate(declaration.PositionAtDeclaration.Latitude, declaration.PositionAtDeclaration.Longitude);
-                        CoordinateSharp.Coordinate goalCoordinate = new CoordinateSharp.Coordinate(declaration.DeclaredGoal.Latitude, declaration.DeclaredGoal.Longitude);
-                        string line = string.Join(',', declaration.GoalNumber, timeStamp, declaration.PositionAtDeclaration.Latitude,
-                            declaration.PositionAtDeclaration.Longitude, posCoordinate.Latitude.Display, posCoordinate.Longitude.Display,
-                            posCoordinate.UTM.LongZone + posCoordinate.UTM.LatZone, posCoordinate.UTM.Easting, posCoordinate.UTM.Northing,
-                            declaration.PositionAtDeclaration.AltitudeGPS, declaration.PositionAtDeclaration.AltitudeBarometric,
-                            declaration.DeclaredGoal.Latitude, declaration.DeclaredGoal.Longitude, goalCoordinate.Latitude.Display,
-                            goalCoordinate.Longitude.Display, goalCoordinate.UTM.LongZone + goalCoordinate.UTM.LatZone, goalCoordinate.UTM.Easting,
-                            goalCoordinate.UTM.Northing, declaration.DeclaredGoal.AltitudeGPS, declaration.DeclaredGoal.AltitudeBarometric);
-                        writer.WriteLine(line);
-                    }
+                    string timeStamp = marker.MarkerLocation.TimeStamp.ToString("dd.MM.yyyy HH:mm:ss");
+                    CoordinateSharp.Coordinate coordinateSharp = new(marker.MarkerLocation.Latitude, marker.MarkerLocation.Longitude);
+                    string line = string.Join(',', marker.MarkerNumber, timeStamp, marker.MarkerLocation.Latitude, marker.MarkerLocation.Longitude,
+                        coordinateSharp.Latitude.Display, coordinateSharp.Longitude.Display, coordinateSharp.UTM.LongZone + coordinateSharp.UTM.LatZone,
+                        coordinateSharp.UTM.Easting, coordinateSharp.UTM.Northing, marker.MarkerLocation.AltitudeGPS, marker.MarkerLocation.AltitudeBarometric);
+                    writer.WriteLine(line);
+                }
+                writer.WriteLine();
+                writer.WriteLine();
+                writer.WriteLine();
+                string declarationHeader = "Declaration Number,Time Stamp,Position Latitude[dec],Position Longitude[dec],Position Latitude[deg],Position Longitude[deg],Position Altitude GPS[m],Position Altitude Barometric[m],Position UTM Zone,Position Easting,Position Norting,Goal Latitude[dec],Goal Longitude[dec],Goal Latitude[deg],Goal Longitude[deg],Goal UTM Zone,Goal Easting,Goal Norting,Goal Altitude GPS m],Goal Altitude Barometric[m]";
+                writer.WriteLine(declarationHeader);
+                foreach (Declaration declaration in track.Declarations)
+                {
+                    string timeStamp = declaration.PositionAtDeclaration.TimeStamp.ToString("dd.MM.yyyy HH:mm:ss");
+                    CoordinateSharp.Coordinate posCoordinate = new(declaration.PositionAtDeclaration.Latitude, declaration.PositionAtDeclaration.Longitude);
+                    CoordinateSharp.Coordinate goalCoordinate = new(declaration.DeclaredGoal.Latitude, declaration.DeclaredGoal.Longitude);
+                    string line = string.Join(',', declaration.GoalNumber, timeStamp, declaration.PositionAtDeclaration.Latitude,
+                        declaration.PositionAtDeclaration.Longitude, posCoordinate.Latitude.Display, posCoordinate.Longitude.Display,
+                        posCoordinate.UTM.LongZone + posCoordinate.UTM.LatZone, posCoordinate.UTM.Easting, posCoordinate.UTM.Northing,
+                        declaration.PositionAtDeclaration.AltitudeGPS, declaration.PositionAtDeclaration.AltitudeBarometric,
+                        declaration.DeclaredGoal.Latitude, declaration.DeclaredGoal.Longitude, goalCoordinate.Latitude.Display,
+                        goalCoordinate.Longitude.Display, goalCoordinate.UTM.LongZone + goalCoordinate.UTM.LatZone, goalCoordinate.UTM.Easting,
+                        goalCoordinate.UTM.Northing, declaration.DeclaredGoal.AltitudeGPS, declaration.DeclaredGoal.AltitudeBarometric);
+                    writer.WriteLine(line);
                 }
             }
         }
 
         private void btImportTask_Click(object sender, EventArgs e)
         {
-            OpenFileDialog openFileDialog = new OpenFileDialog();
-            openFileDialog.DefaultExt = "json";
-            openFileDialog.Filter = "json files(*.json) | *.json";
-            openFileDialog.FilterIndex = 1;
+            OpenFileDialog openFileDialog = new()
+            {
+                DefaultExt = "json",
+                Filter = "json files(*.json) | *.json",
+                FilterIndex = 1
+            };
             if (!string.IsNullOrWhiteSpace(tbCompetitionFolder.Text))
                 openFileDialog.InitialDirectory = tbCompetitionFolder.Text;
             if (openFileDialog.ShowDialog() == DialogResult.OK)
@@ -492,7 +487,7 @@ namespace BalloonTrackAnalyze
                         string jsonString = File.ReadAllText(openFileDialog.FileName);
                         //DonutTask donut = JsonSerializer.Deserialize<DonutTask>(jsonString);
                         DonutTask donut = JsonConvert.DeserializeObject<DonutTask>(jsonString, new JsonSerializerSettings() { TypeNameHandling = TypeNameHandling.Auto });
-                        DonutControl donutControl = new DonutControl(donut);
+                        DonutControl donutControl = new(donut);
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         donutControl.Location = TaskControlLocation;
@@ -506,7 +501,7 @@ namespace BalloonTrackAnalyze
                         string jsonString = File.ReadAllText(openFileDialog.FileName);
                         //PieTask pie = JsonSerializer.Deserialize<PieTask>(jsonString);
                         PieTask pie = JsonConvert.DeserializeObject<PieTask>(jsonString, new JsonSerializerSettings() { TypeNameHandling = TypeNameHandling.Auto });
-                        PieControl pieControl = new PieControl(pie);
+                        PieControl pieControl = new(pie);
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         pieControl.Location = TaskControlLocation;
@@ -520,7 +515,7 @@ namespace BalloonTrackAnalyze
                         string jsonString = File.ReadAllText(openFileDialog.FileName);
                         //ElbowTask elbow = JsonSerializer.Deserialize<ElbowTask>(jsonString);
                         ElbowTask elbow = JsonConvert.DeserializeObject<ElbowTask>(jsonString, new JsonSerializerSettings() { TypeNameHandling = TypeNameHandling.Auto });
-                        ElbowControl elbowControl = new ElbowControl(elbow);
+                        ElbowControl elbowControl = new(elbow);
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         elbowControl.Location = TaskControlLocation;
@@ -534,7 +529,7 @@ namespace BalloonTrackAnalyze
                         string jsonString = File.ReadAllText(openFileDialog.FileName);
                         //LandRunTask landRun = JsonSerializer.Deserialize<LandRunTask>(jsonString);
                         LandRunTask landRun = JsonConvert.DeserializeObject<LandRunTask>(jsonString, new JsonSerializerSettings() { TypeNameHandling = TypeNameHandling.Auto });
-                        LandRunControl landrunControl = new LandRunControl(landRun);
+                        LandRunControl landrunControl = new(landRun);
                         SuspendLayout();
                         plUserControl.Controls.Remove(plUserControl.Controls["taskControl"]);
                         landrunControl.Location = TaskControlLocation;
@@ -545,14 +540,14 @@ namespace BalloonTrackAnalyze
                     }
                     else
                     {
-                        Log(LogSeverityType.Error, "Failed import task: Unkown task type");
+                        Logger?.LogError("Failed import task: Unkown task type");
                         return;
                     }
                 }
                 catch (Exception ex)
                 {
 
-                    Log(LogSeverityType.Error, $"Failed to import task: {ex.Message}");
+                    Logger?.LogError(ex, "Failed to import task");
                     return;
                 }
             }
@@ -564,13 +559,12 @@ namespace BalloonTrackAnalyze
             {
                 if (string.IsNullOrWhiteSpace(tbFlightNumber.Text))
                 {
-                    Log(LogSeverityType.Error, "Please enter a Flight No first");
+                    Logger?.LogError("Failed to create default folder structure: Please enter a Flight No first");
                     return;
                 }
-                int flightNumber;
-                if (!int.TryParse(tbFlightNumber.Text, out flightNumber))
+                if (!int.TryParse(tbFlightNumber.Text, out int flightNumber))
                 {
-                    Log(LogSeverityType.Error, $"Failed to parse Flight No '{tbFlightNumber.Text}' as integer");
+                    Logger?.LogError("Failed to create default folder structure: Failed to parse Flight No '{flightNumber}' as integer", tbFlightNumber.Text);
                     return;
                 }
                 CreateFolderStructure(flightNumber);
