@@ -1,24 +1,27 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Drawing;
-using System.Data;
-using System.Text;
-using System.Windows.Forms;
+﻿using BalloonTrackAnalyze.TaskControls;
 using Competition;
-using LoggerComponent;
 using Coordinates;
+using LoggingConnector;
+using Microsoft.Extensions.Logging;
+using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Windows.Forms;
 
 namespace BalloonTrackAnalyze.ValidationControls
 {
     public partial class GoalToOtherGoalsDistanceRuleControl : UserControl
     {
         #region Properties
+
+        private readonly ILogger<GoalToOtherGoalsDistanceRuleControl> Logger = LogConnector.LoggerFactory.CreateLogger<GoalToOtherGoalsDistanceRuleControl>();
         /// <summary>
         /// The rule to be created or modified with this control
         /// </summary>
-        public GoalToOtherGoalsDistanceRule GoalToOtherGoalsDistanceRule { get; private set; }
+        public GoalToOtherGoalsDistanceRule GoalToOtherGoalsDistanceRule
+        {
+            get; private set;
+        }
 
         /// <summary>
         /// Delegate for the DataValid event
@@ -40,6 +43,7 @@ namespace BalloonTrackAnalyze.ValidationControls
         {
             InitializeComponent();
             btCreate.Text = "Create rule";
+
         }
 
         /// <summary>
@@ -96,18 +100,17 @@ namespace BalloonTrackAnalyze.ValidationControls
         private void btCreate_Click(object sender, EventArgs e)
         {
             bool isDataValid = true;
-            string functionErrorMessage = "Failed to create/modify goal to other goals distance rule: ";
             double minimumDistance = double.NaN;
             if (!string.IsNullOrWhiteSpace(tbMinimumDistance.Text))
             {
                 if (!double.TryParse(tbMinimumDistance.Text, out minimumDistance))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse Min. Distance '{tbMinimumDistance.Text}' as double");
+                    Logger?.LogError("Failed to create/modify goal to other goals distance rule: failed to parse Min. Distance '{minimumDistance}' as double", tbMinimumDistance.Text);
                     isDataValid = false;
                 }
                 if (minimumDistance < 0)
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Min. Distance '{minimumDistance}' must be greater than zero");
+                    Logger?.LogError("Failed to create/modify goal to other goals distance rule: Min. Distance must be greater than zero");
                     isDataValid = false;
                 }
 
@@ -120,12 +123,12 @@ namespace BalloonTrackAnalyze.ValidationControls
             {
                 if (!double.TryParse(tbMaximumDistance.Text, out maximumDistance))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse Max. Distance '{tbMinimumDistance.Text}' as double");
+                    Logger?.LogError("Failed to create/modify goal to other goals distance rule: failed to parse Max. Distance '{tbMinimumDistance.Text}' as double", tbMinimumDistance.Text);
                     isDataValid = false;
                 }
                 if (maximumDistance < 0)
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Max. Distance '{maximumDistance}' must be greater than zero");
+                    Logger?.LogError("Failed to create/modify goal to other goals distance rule: Max. Distance must be greater than zero");
                     isDataValid = false;
                 }
 
@@ -137,15 +140,15 @@ namespace BalloonTrackAnalyze.ValidationControls
             {
                 if (minimumDistance >= maximumDistance)
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Min. Distance '{minimumDistance}[m]' must be smaller than Max. Distance '{maximumDistance}[m]'");
+                    Logger?.LogError("Failed to create/modify goal to other goals distance rule: Min. Distance '{minimumDistance}[m]' must be smaller than Max. Distance '{maximumDistance}[m]'", minimumDistance, maximumDistance);
                     isDataValid = false;
                 }
             }
 
-            List<int> goalNumbers = new List<int>();
+            List<int> goalNumbers = [];
             if (!string.IsNullOrWhiteSpace(tbGoalNumbers.Text))
-                if (tbGoalNumbers.Text.ToLowerInvariant() != "all")
-                    goalNumbers = Array.ConvertAll(tbGoalNumbers.Text.Split(','), int.Parse).ToList();
+                if (!tbGoalNumbers.Text.Equals("all", StringComparison.InvariantCultureIgnoreCase))
+                    goalNumbers = [.. Array.ConvertAll(tbGoalNumbers.Text.Split(','), int.Parse)];
 
 
 
@@ -166,16 +169,6 @@ namespace BalloonTrackAnalyze.ValidationControls
         protected virtual void OnDataValid()
         {
             DataValid?.Invoke();
-        }
-
-        /// <summary>
-        /// Logs a user message
-        /// </summary>
-        /// <param name="logSeverity">the severity of the message</param>
-        /// <param name="text">the message text</param>
-        private void Log(LogSeverityType logSeverity, string text)
-        {
-            Logger.Log(this, logSeverity, text);
         }
         #endregion
     }
