@@ -1,14 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Drawing;
-using System.Data;
-using System.Text;
-using System.Windows.Forms;
-using Coordinates;
-using LoggerComponent;
+﻿using BalloonTrackAnalyze.TaskControls;
 using Competition;
+using Coordinates;
+using LoggingConnector;
+using Microsoft.Extensions.Logging;
+using System;
+using System.Collections.Generic;
 using System.Linq;
+using System.Windows.Forms;
 
 namespace BalloonTrackAnalyze.ValidationControls
 {
@@ -16,10 +14,14 @@ namespace BalloonTrackAnalyze.ValidationControls
     {
         #region Properties
 
+        private readonly ILogger<MarkerToOtherMarkersDistanceRuleControl> Logger = LogConnector.LoggerFactory.CreateLogger<MarkerToOtherMarkersDistanceRuleControl>();
         /// <summary>
         /// The rule to be created or modified with this control
         /// </summary>
-        public MarkerToOtherMarkersDistanceRule MarkerToOtherMarkersDistanceRule { get; private set; }
+        public MarkerToOtherMarkersDistanceRule MarkerToOtherMarkersDistanceRule
+        {
+            get; private set;
+        }
 
         /// <summary>
         /// Delegate for the DataValid event
@@ -31,6 +33,7 @@ namespace BalloonTrackAnalyze.ValidationControls
         /// </summary>
         public event DataValidDelegate DataValid;
         #endregion
+
         #region Constructors
         /// <summary>
         /// Default constructor
@@ -96,18 +99,17 @@ namespace BalloonTrackAnalyze.ValidationControls
         private void btCreate_Click(object sender, EventArgs e)
         {
             bool isDataValid = true;
-            string functionErrorMessage = "Failed to create/modify marker to other markers distance rule: ";
             double minimumDistance = double.NaN;
             if (!string.IsNullOrWhiteSpace(tbMinimumDistance.Text))
             {
                 if (!double.TryParse(tbMinimumDistance.Text, out minimumDistance))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse Min. Distance '{tbMinimumDistance.Text}' as double");
+                    Logger?.LogError("Failed to create/modify marker to other markers distance rule: failed to parse Min. Distance '{tbMinimumDistance.Text}' as double", tbMinimumDistance.Text);
                     isDataValid = false;
                 }
                 if (minimumDistance < 0)
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Min. Distance '{minimumDistance}' must be greater than zero");
+                    Logger?.LogError("Failed to create/modify marker to other markers distance rule: Min. Distance '{minimumDistance}' must be greater than zero", minimumDistance);
                     isDataValid = false;
                 }
 
@@ -120,12 +122,12 @@ namespace BalloonTrackAnalyze.ValidationControls
             {
                 if (!double.TryParse(tbMaximumDistance.Text, out maximumDistance))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse Max. Distance '{tbMinimumDistance.Text}' as double");
+                    Logger?.LogError("Failed to create/modify marker to other markers distance rule: Failed to parse Max. Distance '{tbMinimumDistance.Text}' as double", tbMaximumDistance.Text);
                     isDataValid = false;
                 }
                 if (maximumDistance < 0)
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Max. Distance '{maximumDistance}' must be greater than zero");
+                    Logger?.LogError("Failed to create/modify marker to other markers distance rule: Max. Distance '{maximumDistance}' must be greater than zero", maximumDistance);
                     isDataValid = false;
                 }
 
@@ -137,14 +139,14 @@ namespace BalloonTrackAnalyze.ValidationControls
             {
                 if (minimumDistance >= maximumDistance)
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Min. Distance '{minimumDistance}[m]' must be smaller than Max. Distance '{maximumDistance}[m]'");
+                    Logger?.LogError("Failed to create/modify marker to other markers distance rule: Min. Distance '{minimumDistance}[m]' must be smaller than Max. Distance '{maximumDistance}[m]'", minimumDistance, maximumDistance);
                     isDataValid = false;
                 }
             }
-            List<int> markerNumbers = new List<int>();
+            List<int> markerNumbers = [];
             if (!string.IsNullOrWhiteSpace(tbMarkerNumbers.Text))
-                if (tbMarkerNumbers.Text.ToLowerInvariant() != "all")
-                    markerNumbers = Array.ConvertAll(tbMarkerNumbers.Text.Split(','), int.Parse).ToList();
+                if (!tbMarkerNumbers.Text.Equals("all", StringComparison.InvariantCultureIgnoreCase))
+                    markerNumbers = [.. Array.ConvertAll(tbMarkerNumbers.Text.Split(','), int.Parse)];
 
 
             if (isDataValid)
@@ -166,15 +168,6 @@ namespace BalloonTrackAnalyze.ValidationControls
             DataValid?.Invoke();
         }
 
-        /// <summary>
-        /// Logs a user message
-        /// </summary>
-        /// <param name="logSeverity">the severity of the message</param>
-        /// <param name="text">the message text</param>
-        private void Log(LogSeverityType logSeverity, string text)
-        {
-            Logger.Log(this, logSeverity, text);
-        }
         #endregion
     }
 }
