@@ -1,11 +1,10 @@
 ﻿using Competition.Validation;
 using Coordinates;
-using LoggerComponent;
-using System;
+using LoggingConnector;
+using Microsoft.Extensions.Logging;
+using Newtonsoft.Json;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
-using System.Text;
 
 namespace Competition
 {
@@ -16,6 +15,8 @@ namespace Competition
         public class PieTier
         {
             #region Properties
+            [JsonIgnore()]
+            private readonly ILogger<PieTier> Logger = LogConnector.LoggerFactory.CreateLogger<PieTier>();
 
             /// <summary>
             /// The target goal number
@@ -30,7 +31,7 @@ namespace Competition
             /// Specify whether or not re-entrance in the donut is allowed
             /// <para>mandatory</para>
             /// </summary>
-            public bool IsReentranceAllowed
+            public bool IsReEntranceAllowed
             {
                 get; set;
             } = true;
@@ -77,7 +78,7 @@ namespace Competition
             public List<IDeclarationValidationRules> DeclarationValidationRules
             {
                 get; set;
-            } = new List<IDeclarationValidationRules>();
+            } = [];
 
             public PieTier()
             {
@@ -99,13 +100,12 @@ namespace Competition
             {
                 string functionErrorMessage = $"Failed to calculate result for {this} and Pilot '#{track.Pilot.PilotNumber}{(!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : "")}': ";
                 result = 0.0;
-                List<(int, Coordinate)> trackPointsInTier = new List<(int trackPointNumber, Coordinate coordinate)>();
+                List<(int, Coordinate)> trackPointsInTier = [];
 
                 Declaration targetDeclaration = ValidationHelper.GetValidDeclaration(track, GoalNumber, DeclarationValidationRules);
                 if (targetDeclaration == null)
                 {
-                    //Debug.WriteLine("No valid goal found");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"No valid goal found for goal '#{GoalNumber}'");
+                    Logger?.LogError("Failed to calculate result for '{task}' and Pilot '#{pilotNumber}{pilotName}': No valid goal found for goal '#{goalNumber}'", ToString(), track.Pilot.PilotNumber, (!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : ""), GoalNumber);
                     return false;
                 }
                 List<Coordinate> coordinates = track.TrackPoints;
@@ -128,12 +128,12 @@ namespace Competition
                 {
                     double distanceToGoal = CoordinateHelpers.Calculate2DDistanceHavercos(coordinates[index], targetDeclaration.DeclaredGoal);//calculate distance to goal
                     if (distanceToGoal <= Radius)//save all trackpoints within the radius
-                        trackPointsInTier.Add((track.TrackPoints.FindIndex(x=>x==coordinates[index]), coordinates[index]));
+                        trackPointsInTier.Add((track.TrackPoints.FindIndex(x => x == coordinates[index]), coordinates[index]));
 
                 }
-                List<List<Coordinate>> chunksInTier = new List<List<Coordinate>>();
+                List<List<Coordinate>> chunksInTier = [];
                 int addIndex = 0;
-                chunksInTier.Add(new List<Coordinate>());
+                chunksInTier.Add([]);
                 for (int index = 0; index < trackPointsInTier.Count - 1; index++)
                 {
                     if (trackPointsInTier[index + 1].Item1 - trackPointsInTier[index].Item1 == 1)//trackpoints are successive
@@ -150,12 +150,12 @@ namespace Competition
                     }
                     else//trackpoints are not successive -> create new chunk
                     {
-                        chunksInTier.Add(new List<Coordinate>());
+                        chunksInTier.Add([]);
                         addIndex++;
                     }
                 }
 
-                if (!IsReentranceAllowed)//evaluate first chunk only
+                if (!IsReEntranceAllowed)//evaluate first chunk only
                 {
                     if (chunksInTier[0].Count >= 2)
                     {
@@ -193,13 +193,13 @@ namespace Competition
             /// <param name="radius">The radius of the pie tier in meter (mandatory)</param>
             /// <param name="lowerBoundary">Lower boundary of the donut in meter (optional; use double.NaN to omit)</param>
             /// <param name="upperBoundary">Upper boundary of the donut in meter (optional; use double.NaN to omit)</param>
-            /// <param name="isReentranceAllowed">Specify whether or not re-entrance in the donut is allowed (mandatory)</param>
+            /// <param name="isReEntranceAllowed">Specify whether or not re-entrance in the donut is allowed (mandatory)</param>
             /// <param name="declarationValidationRules">List of rules for declaration validation (optional; leave list empty to omit)</param>
-            public void SetupPieTier(int goalNumber, double radius, bool isReentranceAllowed, double multiplier, double lowerBoundary, double upperBoundary, List<IDeclarationValidationRules> declarationValidationRules)
+            public void SetupPieTier(int goalNumber, double radius, bool isReEntranceAllowed, double multiplier, double lowerBoundary, double upperBoundary, List<IDeclarationValidationRules> declarationValidationRules)
             {
                 GoalNumber = goalNumber;
                 Radius = radius;
-                IsReentranceAllowed = isReentranceAllowed;
+                IsReEntranceAllowed = isReEntranceAllowed;
                 Multiplier = multiplier;
                 LowerBoundary = lowerBoundary;
                 UpperBoundary = upperBoundary;
@@ -213,16 +213,14 @@ namespace Competition
             #endregion
 
             #region Private methods
-            private void Log(LogSeverityType logSeverity, string text)
-            {
-                Logger.Log(this, logSeverity, text);
-            }
+
             #endregion
         }
         #endregion
 
         #region Properties
 
+        private readonly ILogger<PieTask> Logger = LogConnector.LoggerFactory.CreateLogger<PieTask>();
         /// <summary>
         /// The task number
         /// <para>mandatory</para>
@@ -239,7 +237,7 @@ namespace Competition
         public List<PieTier> Tiers
         {
             get; set;
-        } = new List<PieTier>();
+        } = [];
         #endregion
 
         public PieTask()
@@ -258,14 +256,12 @@ namespace Competition
         /// <returns>true:success;false:error</returns>
         public bool CalculateResults(Track track, bool useGPSAltitude, out double result)
         {
-            string functionErrorMessage = $"Failed to calculate result for {this} and Pilot '#{track.Pilot.PilotNumber}{(!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : "")}': ";
             result = 0.0;
             foreach (PieTier tier in Tiers)
             {
-                double tempResult;
-                if (!tier.CalculateTierResult(track, useGPSAltitude, out tempResult))
+                if (!tier.CalculateTierResult(track, useGPSAltitude, out double tempResult))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + "Failed to calculate tier result");
+                    Logger?.LogError("Failed to calculate result for '{task}' and Pilot '#{pilotNumber}{pilotName}': Failed to calculate tier result", ToString(), track.Pilot.PilotNumber, (!string.IsNullOrWhiteSpace(track.Pilot.FirstName) ? $"({track.Pilot.FirstName},{track.Pilot.LastName})" : ""));
                     return false;
                 }
                 result += tempResult;
@@ -290,10 +286,6 @@ namespace Competition
         #endregion
 
         #region Private methods
-        private void Log(LogSeverityType logSeverity, string text)
-        {
-            Logger.Log(this, logSeverity, text);
-        }
         #endregion
     }
 
