@@ -1,14 +1,17 @@
 ﻿using Coordinates;
-using LoggerComponent;
-using System;
+using LoggingConnector;
+using Microsoft.Extensions.Logging;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 
 namespace Competition.Validation
 {
     public static class ValidationHelper
     {
+
+        private static readonly ILogger Logger = LogConnector.LoggerFactory.CreateLogger(nameof(ValidationHelper));
+
+
         /// <summary>
         /// Applies all specified declaration rules for goals of the specified goal number and return the latest valid declaration
         /// </summary>
@@ -19,10 +22,10 @@ namespace Competition.Validation
         public static Declaration GetValidDeclaration(Track track, int goalNumber, List<IDeclarationValidationRules> declarationValidationRules)
         {
             List<Declaration> declarations = track.Declarations.Where(x => x.GoalNumber == goalNumber).ToList();
-            List<Declaration> validDeclarations = new List<Declaration>();
+            List<Declaration> validDeclarations = [];
             if (declarations.Count == 0)
             {
-                Logger.Log(LogSeverityType.Warning, $"No declaration found for goal number '{goalNumber}'");
+                Logger?.LogWarning("No declaration found for goal number '{goalNumber}'",goalNumber);
                 return null;
             }
             else
@@ -46,7 +49,7 @@ namespace Competition.Validation
                 }
                 if (validDeclarations.Count == 0)
                 {
-                    Logger.Log(LogSeverityType.Warning, $"No declaration of goal number '{goalNumber}' is conform to specified rules");
+                    Logger?.LogWarning("No declaration of goal number '{goalNumber}' is conform to specified rules", goalNumber);    
                     return null;
                 }
                 else if (validDeclarations.Count == 1)
@@ -86,7 +89,7 @@ namespace Competition.Validation
             if (markerDrop == null)
             {
                 //Console.WriteLine($"No Marker '{FirstMarkerNumber}' found");
-                Logger.Log(LogSeverityType.Warning, $"No Marker '{markerNumber}' found");
+                Logger?.LogWarning("No Marker '{markerNumber}' found", markerNumber);
                 isValid = false;
             }
             else
@@ -100,7 +103,7 @@ namespace Competition.Validation
                 }
                 if (!isValid)
                 {
-                    Logger.Log(LogSeverityType.Warning, $"Marker '{markerNumber}' is not conform to specified rules");
+                    Logger?.LogWarning("Marker '{markerNumber}' is not conform to specified rules", markerNumber);
                 }
             }
             return isValid;
