@@ -1,17 +1,42 @@
-﻿using System;
+﻿using LoggingConnector;
+using Microsoft.Extensions.Logging;
+using System;
 using System.Collections.Generic;
-using System.Text;
 using System.IO;
-using System.Diagnostics;
 using System.Linq;
-using System.Text.RegularExpressions;
-using LoggerComponent;
 
 
 namespace Coordinates.Parsers
 {
     public static class BalloonLiveParser
     {
+        private static ILogger Logger = LogConnector.LoggerFactory.CreateLogger(nameof(BalloonLiveParser));
+
+
+        private static bool MarkerDrop_HasAdditionalLatitudeDecimals = false;
+        private static int MarkerDrop_StartOfAdditionalLatitudeDecimals = -1;
+        private static int MarkerDrop_EndOfAdditionalLatitudeDecimals = -1;
+
+        private static bool MarkerDrop_HasAdditionalLongitudeDecimals = false;
+        private static int MarkerDrop_StartOfAdditionalLongitudeDecimals = -1;
+        private static int MarkerDrop_EndOfAdditionalLongitudeDecimals = -1;
+
+        private static bool Declaration_HasAdditionalLatitudeDecimals = false;
+        private static int Declaration_StartOfAdditionalLatitudeDecimals = -1;
+        private static int Declaration_EndOfAdditionalLatitudeDecimals = -1;
+
+        private static bool Declaration_HasAdditionalLongitudeDecimals = false;
+        private static int Declaration_StartOfAdditionalLongitudeDecimals = -1;
+        private static int Declaration_EndOfAdditionalLongitudeDecimals = -1;
+
+        private static bool TrackPoint_HasAdditionalLatitudeDecimals = false;
+        private static int TrackPoint_StartOfAdditionalLatitudeDecimals = -1;
+        private static int TrackPoint_EndOfAdditionalLatitudeDecimals = -1;
+
+        private static bool TrackPoint_HasAdditionalLongitudeDecimals = false;
+        private static int TrackPoint_StartOfAdditionalLongitudeDecimals = -1;
+        private static int TrackPoint_EndOfAdditionalLongitudeDecimals = -1;
+
         #region API
 
         /// <summary>
@@ -19,38 +44,37 @@ namespace Coordinates.Parsers
         /// </summary>
         /// <param name="fileNameAndPath">the file path and name of the .igc file</param>
         /// <param name="track">output parameter. the parsed track from the file</param>
-        /// <param name="defaultGoalAlitude">the default declaration altitude in meters to be used when the pilot didn't declared one</param>
+        /// <param name="referenceCoordinate">provide a reference coordinate to be used complete the missing information of a goal declaration
+        /// <para>a goal declaration not in the zone 6/7 format is ambiguous</para>
+        /// <para>this is an optional parameter, the parse will use either marker drop 1 or the position at declaration if not reference point has been provided</para></param>
         /// <returns>true:success; false:error</returns>
-        public static bool ParseFile(string fileNameAndPath, out Track track, double defaultGoalAlitude=0.0)
+        public static bool ParseFile(string fileNameAndPath, out Track track, Coordinate referenceCoordinate = null)
         {
-            string functionErrorMessage = $"Failed to parse file '{fileNameAndPath}':";
+
             track = null;
             try
             {
-
-                FileInfo fileInfo = new FileInfo(fileNameAndPath);
+                Logger?.LogInformation("Parsing file '{filePathAndName}'", fileNameAndPath);
+                FileInfo fileInfo = new(fileNameAndPath);
                 if (!fileInfo.Exists)
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"The file '{fileNameAndPath}' does not exists");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"The file '{fileNameAndPath}' does not exists");
+                    Logger?.LogError("Failed to parse the file '{filePathAndName}': the file does not exists", fileNameAndPath);
                     return false;
                 }
 
                 if (!fileInfo.Extension.EndsWith("igc"))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"The file extension '{fileInfo.Extension}' is not supported");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"The file extension '{fileInfo.Extension}' is not supported");
+                    Logger?.LogError("Failed to parse the file '{filePathAndName}': the file extension '{fileExtension}' is not supported", fileNameAndPath, fileInfo.Extension);
                     return false;
                 }
+                ResetProperties();
                 track = new Track();
                 int pilotNumber = -1;
                 string pilotIdentifier = "";
-                DateTime date = new DateTime();
-                //int goalNortingDigits = -1;
-                //int goalEastingDigits = -1;
+                DateTime date = new();
                 bool declaredAltitudeIsInFeet = true;
-                List<string> lines = new List<string>();
-                using (StreamReader reader = new StreamReader(fileNameAndPath))
+                List<string> lines = [];
+                using (StreamReader reader = new(fileNameAndPath))
                 {
                     while (!reader.EndOfStream)
                     {
@@ -58,20 +82,18 @@ namespace Coordinates.Parsers
                         lines.Add(line);
                     }
                 }
-                //string[] identifierLine = lines.Where(x => x.StartsWith("AXXX")).ToArray();
                 string[] identifierLine = lines.Where(x => x.StartsWith("AXBL")).ToArray();
                 if (identifierLine.Length == 0)
                 {
                     identifierLine = lines.Where(x => x.StartsWith("AXXX")).ToArray();
                     if (identifierLine.Length == 0)
                     {
-                        Log(LogSeverityType.Error, "Line with Pilot Identifier 'AXBL' or 'AXXX' is missing");
+                        Logger?.LogError("Failed to parse the file '{filePathAndName}': Line with Pilot Identifier 'AXBL' or 'AXXX' is missing", fileNameAndPath);
                         return false;
                     }
                     if (identifierLine.Length > 1)
                     {
-                        //Debug.WriteLine("More the one line with Pilot Identifier is found. First occurrence will be used.");
-                        Log(LogSeverityType.Warning, "More the one line with Pilot Identifier ('AXXX') is found. First occurrence will be used.");
+                        Logger?.LogWarning("More the one line with Pilot Identifier ('AXXX') were found. First occurrence will be used.");
                     }
                     else
                     {
@@ -80,32 +102,18 @@ namespace Coordinates.Parsers
                 }
                 if (identifierLine.Length > 1)
                 {
-                    //Debug.WriteLine("More the one line with Pilot Identifier is found. First occurrence will be used.");
-                    Log(LogSeverityType.Warning, "More the one line with Pilot Identifier ('AXBL') is found. First occurrence will be used.");
+                    Logger?.LogWarning("More the one line with Pilot Identifier ('AXBL') were found. First occurrence will be used.");
                 }
                 else
                 {
                     pilotIdentifier = identifierLine[0][4..12];
                 }
-                //pilotIdentifier = identifierLine[0].Replace("AXXX", "").Replace("BalloonLive", "");
 
 
                 string[] headerLines = lines.Where(x => x.StartsWith('H')).ToArray();
-                bool markerDrop_HasAdditionalLatitudeDecimals;
-                int markerDrop_StartOfAdditionalLatitudeDecimals;
-                int markerDrop_EndOfAdditionalLatitudeDecimals;
-                bool markerDrop_HasAdditionalLongitudeDecimals;
-                int markerDrop_StartOfAdditionalLongitudeDecimals;
-                int markerDrop_EndOfAdditionalLongitudeDecimals;
-                bool declaration_HasAdditionalLatitudeDecimals;
-                int declaration_StartOfAdditionalLatitudeDecimals;
-                int declaration_EndOfAdditionalLatitudeDecimals;
-                bool declaration_HasAdditionalLongitudeDecimals;
-                int declaration_StartOfAdditionalLongitudeDecimals;
-                int declaration_EndOfAdditionalLongitudeDecimals;
-                if (!ParseHeaders(headerLines, out pilotNumber, out date, out markerDrop_HasAdditionalLatitudeDecimals, out markerDrop_StartOfAdditionalLatitudeDecimals, out markerDrop_EndOfAdditionalLatitudeDecimals, out markerDrop_HasAdditionalLongitudeDecimals, out markerDrop_StartOfAdditionalLongitudeDecimals, out markerDrop_EndOfAdditionalLongitudeDecimals, out declaration_HasAdditionalLatitudeDecimals, out declaration_StartOfAdditionalLatitudeDecimals, out declaration_EndOfAdditionalLatitudeDecimals, out declaration_HasAdditionalLongitudeDecimals, out declaration_StartOfAdditionalLongitudeDecimals, out declaration_EndOfAdditionalLongitudeDecimals))
+                if (!ParseHeaders(headerLines, out pilotNumber, out date))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage);
+                    Logger?.LogError("Failed to parse the file '{filePathAndName}': Failed to parse header lines", fileNameAndPath);
                     return false;
                 }
 
@@ -126,109 +134,83 @@ namespace Coordinates.Parsers
                     if (configLine.StartsWith("LXXX BLSSN"))
                     {
                         string sensBoxSerialNumber = configLine.Split('=').Last();
-                        if (!track.AdditionalPropertiesFromIGCFile.ContainsKey("SensBoxSerialNumber"))
-                            track.AdditionalPropertiesFromIGCFile.Add("SensBoxSerialNumber", sensBoxSerialNumber);
-                    }
-                    if (configLine.StartsWith("LXXX loggerinterval"))
-                    {
-                        string loggerInterval = configLine.Split("=").Last();
-                        if (!track.AdditionalPropertiesFromIGCFile.ContainsKey("LoggerInterval"))
-                            track.AdditionalPropertiesFromIGCFile.Add("LoggerInterval", loggerInterval.Replace("s",""));
+                        if (track.AdditionalPropertiesFromIGCFile.TryAdd("SensBoxSerialNumber", sensBoxSerialNumber))
+                        {
+                            Logger?.LogWarning("Failed to add SensBoxSerialNumber '{SensBoxSerialNumber}' to the track", sensBoxSerialNumber);
+                        }
                     }
                 }
 
                 string iRecordLine = lines.Where(x => x.StartsWith('I')).FirstOrDefault();
-                int offset = 3;
-                bool trackPoint_HasAdditionalLatitudeDecimals = false;
-                int trackPoint_StartOfAdditionalLatitudeDecimals = -1;
-                int trackPoint_EndOfAdditionalLatitudeDecimals = -1;
-                bool trackPoint_HasAdditionalLongitudeDecimals = false;
-                int trackPoint_StartOfAdditionalLongitudeDecimals = -1;
-                int trackPoint_EndOfAdditionalLongitudeDecimals = -1;
                 if (!string.IsNullOrWhiteSpace(iRecordLine))
                 {
-                    int numberOfAdditions;
-                    if (!int.TryParse(iRecordLine[1..3], out numberOfAdditions))
+                    if (!int.TryParse(iRecordLine[1..3], out int numberOfAdditions))
                     {
-                        Log(LogSeverityType.Error, functionErrorMessage + " Failed to parse number of additions from I-record");
+
+                        Logger?.LogError("Failed to parse the file '{filePathAndName}': Failed to parse number of additions from I-record", fileNameAndPath);
                         return false;
                     }
-
+                    int offset = 3;
                     for (int index = 0; index < numberOfAdditions; index++)
                     {
                         offset = index * 7 + 3;
-                        int startPosition;
-                        if (!int.TryParse(iRecordLine[offset..(offset + 2)], out startPosition))
+                        if (!int.TryParse(iRecordLine[offset..(offset + 2)], out int startPosition))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse start position of addition no {index + 1}");
+                            Logger?.LogError("Failed to parse the file '{filePathAndName}': Failed to parse start position of I-record addition no {additionNumber}", fileNameAndPath, index + 1);
                             return false;
                         }
                         offset += 2;
-                        int stopPosition;
-                        if (!int.TryParse(iRecordLine[(offset)..(offset + 2)], out stopPosition))
+                        if (!int.TryParse(iRecordLine[(offset)..(offset + 2)], out int stopPosition))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse stop position of addition no {index + 1}");
+                            Logger?.LogError("Failed to parse the file '{filePathAndName}': Failed to parse stop position of I-record addition no {additionNumber}", fileNameAndPath, index + 1);
                             return false;
                         }
                         offset += 2;
                         string additionIdentifier = iRecordLine[(offset)..(offset + 3)];
 
-                        if (additionIdentifier.ToUpper() == "LAD")
+                        if (additionIdentifier.Equals("LAD", StringComparison.CurrentCultureIgnoreCase))
                         {
-                            trackPoint_HasAdditionalLatitudeDecimals = true;
-                            trackPoint_StartOfAdditionalLatitudeDecimals = startPosition - 1;//adjust to zero based index
-                            trackPoint_EndOfAdditionalLatitudeDecimals = stopPosition;
+                            TrackPoint_HasAdditionalLatitudeDecimals = true;
+                            TrackPoint_StartOfAdditionalLatitudeDecimals = startPosition - 1;//adjust to zero based index
+                            TrackPoint_EndOfAdditionalLatitudeDecimals = stopPosition;
                         }
-                        if (additionIdentifier.ToUpper() == "LOD")
+                        if (additionIdentifier.Equals("LOD", StringComparison.CurrentCultureIgnoreCase))
                         {
-                            trackPoint_HasAdditionalLongitudeDecimals = true;
-                            trackPoint_StartOfAdditionalLongitudeDecimals = startPosition - 1;//adjust to zero based index
-                            trackPoint_EndOfAdditionalLongitudeDecimals = stopPosition;
+                            TrackPoint_HasAdditionalLongitudeDecimals = true;
+                            TrackPoint_StartOfAdditionalLongitudeDecimals = startPosition - 1;//adjust to zero based index
+                            TrackPoint_EndOfAdditionalLongitudeDecimals = stopPosition;
                         }
                     }
                 }
                 string[] positionSourceEvents = lines.Where(x => x.StartsWith('E') && x.Contains("XS")).ToArray();
                 bool isFirstSourceEvent = true;
-                DateTime timeStamp;
-                bool isPrimarySource;
-                bool isBallonLiveSensor;
-                string blsSerialNumber;
                 foreach (string positionSourceEvent in positionSourceEvents)
                 {
-                    if (!ParseSourceEvent(positionSourceEvent, date, out timeStamp, out isPrimarySource, out isBallonLiveSensor, out blsSerialNumber))
+                    if (!ParseSourceEvent(positionSourceEvent, date, out DateTime timeStamp, out bool isPrimarySource, out bool isBallonLiveSensor, out string blsSerialNumber))
                     {
-                        Log(LogSeverityType.Error, functionErrorMessage + "Failed to parse position source event");
+                        Logger?.LogError("Failed to parse the file '{filePathAndName}': Failed to parse position source event", fileNameAndPath);
                         return false;
                     }
                     if (isFirstSourceEvent)
                     {
                         isFirstSourceEvent = false;
-                        Log(LogSeverityType.Info, $"Position source at track start is: '{(isPrimarySource ? "primary" : "fallback")}' position source '{(isBallonLiveSensor ? "Ballon Live Sensor" : "Phone Internal")}' {(isBallonLiveSensor ? $"with serial number '{blsSerialNumber}'" : "")}");
+                        Logger?.LogInformation("Position source of track is: '{sourceType}' position source '{source}' {blsSerialNumber}", (isPrimarySource ? "primary" : "fallback"), (isBallonLiveSensor ? "Ballon Live Sensor" : "Phone Internal"), (isBallonLiveSensor ? $"with serial number '{blsSerialNumber}'" : ""));
                     }
                     else
                     {
-                        Log(LogSeverityType.Warning, $"Caution: Change of position source detected at '{timeStamp:dd-MMM-yyyy HH\\:mm\\:ss}' :  '{(isPrimarySource ? "primary" : "fallback")}' source '{(isBallonLiveSensor ? "Ballon Live Sensor" : "Phone Internal")}' {(isBallonLiveSensor ? $"with serial number '{blsSerialNumber}'" : "")}");
-                        if (!track.AdditionalPropertiesFromIGCFile.ContainsKey("Change of position source"))
-                            track.AdditionalPropertiesFromIGCFile.Add("Change of position source", "yes");
+                        Logger?.LogWarning("Caution, change of position source detected at '{timestamp}': {sourceType}' position source '{source}' {blsSerialNumber}", timeStamp.ToString("dd-MMM-yyyy HH:mm:ss"), (isPrimarySource ? "primary" : "fallback"), (isBallonLiveSensor ? "Ballon Live Sensor" : "Phone Internal"), (isBallonLiveSensor ? $"with serial number '{blsSerialNumber}'" : ""));
+                        if (!track.AdditionalPropertiesFromIGCFile.TryAdd("Change of position source", "yes"))
+                            Logger?.LogWarning("Failed to add 'Change of position source' to the track");
                     }
                 }
-                if (!track.AdditionalPropertiesFromIGCFile.ContainsKey("Change of position source"))
-                    track.AdditionalPropertiesFromIGCFile.Add("Change of position source", "no");
-
+                if (!track.AdditionalPropertiesFromIGCFile.TryAdd("Change of position source", "no"))
+                    Logger?.LogWarning("Failed to add 'Change of position source' to the track");
                 string[] trackPointLines = lines.Where(x => x.StartsWith('B')).ToArray();
                 foreach (string trackPointLine in trackPointLines)
                 {
-                    Coordinate coordinate;
-                    if (!ParseTrackPoint(trackPointLine, date,
-                        trackPoint_HasAdditionalLatitudeDecimals,
-                        trackPoint_StartOfAdditionalLatitudeDecimals,
-                        trackPoint_EndOfAdditionalLatitudeDecimals,
-                        trackPoint_HasAdditionalLongitudeDecimals,
-                        trackPoint_StartOfAdditionalLongitudeDecimals,
-                        trackPoint_EndOfAdditionalLongitudeDecimals, out coordinate))
+                    if (!ParseTrackPoint(trackPointLine, date, out Coordinate coordinate))
                     {
-                        //Debug.WriteLine(functionErrorMessage + "Failed to parse trackpoint");
-                        Log(LogSeverityType.Error, functionErrorMessage + "Failed to parse trackpoint");
+                        Logger?.LogError("Failed to parse the file '{filePathAndName}': Failed to parse trackpoint", fileNameAndPath);
                         return false;
                     }
                     track.TrackPoints.Add(coordinate);
@@ -237,135 +219,83 @@ namespace Coordinates.Parsers
                 string[] markerDropLines = lines.Where(x => x.StartsWith('E') && x.Contains("XX0")).ToArray();
                 foreach (string markerDropLine in markerDropLines)
                 {
-                    MarkerDrop markerDrop;
-                    if (!ParseMarkerDrop(markerDropLine, date,
-                        markerDrop_HasAdditionalLatitudeDecimals,
-                        markerDrop_StartOfAdditionalLatitudeDecimals,
-                        markerDrop_EndOfAdditionalLatitudeDecimals,
-                        markerDrop_HasAdditionalLongitudeDecimals,
-                        markerDrop_StartOfAdditionalLongitudeDecimals,
-                        markerDrop_EndOfAdditionalLongitudeDecimals,
-                        out markerDrop))
+                    if (!ParseMarkerDrop(markerDropLine, date, out MarkerDrop markerDrop))
                     {
-                        //Debug.WriteLine(functionErrorMessage + "Failed to parse marker drop");
-                        Log(LogSeverityType.Error, functionErrorMessage + "Failed to parse marker drop");
+                        Logger?.LogError("Failed to parse the file '{filePathAndName}': Failed to parse marker drop", fileNameAndPath);
                         return false;
                     }
                     track.MarkerDrops.Add(markerDrop);
                 }
-
-                Coordinate referenceCoordinate = null;
-                if (track.MarkerDrops.Count > 0)
+                if (referenceCoordinate is null)
                 {
-                    MarkerDrop markerDrop = track.MarkerDrops.Find(x => x.MarkerNumber == 1);
-                    if (markerDrop != null)
-                        referenceCoordinate = markerDrop.MarkerLocation;
+                    if (track.MarkerDrops.Count > 0)
+                    {
+                        MarkerDrop markerDrop = track.MarkerDrops.Find(x => x.MarkerNumber == 1);
+                        if (markerDrop != null)
+                        {
+                            referenceCoordinate = markerDrop.MarkerLocation;
+                        }
+                        else
+                        {
+                            referenceCoordinate = track.MarkerDrops[0].MarkerLocation;
+                            Logger?.LogWarning("No marker 1 dropped. Marker '{markerNumber}' will be used as goal declaration reference", track.MarkerDrops[0].MarkerNumber);
+                        }
+                    }
                     else
                     {
-                        referenceCoordinate = track.MarkerDrops[0].MarkerLocation;
-                        //Debug.WriteLine($"No marker 1 dropped. Marker '{track.MarkerDrops[0].MarkerNumber}' will be used as goal declaration reference");
-                        Log(LogSeverityType.Warning, $"No marker 1 dropped. Marker '{track.MarkerDrops[0].MarkerNumber}' will be used as goal declaration reference");
+                        Logger?.LogWarning("No marker drops found. Position at declaration will be used as reference instead");
                     }
-                }
-                else
-                {
-                    //Debug.WriteLine("No marker drops found. Position at declaration will be used as reference instead");
-                    Log(LogSeverityType.Warning, "No marker drops found. Position at declaration will be used as reference instead");
                 }
                 string[] goalDeclarationLines = lines.Where(x => x.StartsWith('E') && x.Contains("XL1")).ToArray();
                 foreach (string goalDeclarationLine in goalDeclarationLines)
                 {
 
-                    Declaration declaration;
-                    if (!ParseGoalDeclaration(goalDeclarationLine, 
-                        date, 
-                        declaredAltitudeIsInFeet,
-                        defaultGoalAlitude,
-                        referenceCoordinate,
-                        declaration_HasAdditionalLatitudeDecimals,
-                        declaration_StartOfAdditionalLatitudeDecimals,
-                        declaration_EndOfAdditionalLatitudeDecimals,
-                        declaration_HasAdditionalLongitudeDecimals,
-                        declaration_StartOfAdditionalLongitudeDecimals,
-                        declaration_EndOfAdditionalLongitudeDecimals,
-                        out declaration))
+                    if (!ParseGoalDeclaration(goalDeclarationLine, date, declaredAltitudeIsInFeet, referenceCoordinate, out Declaration declaration))
                     {
-                        //Debug.WriteLine(functionErrorMessage + "Failed to parse goal declaration");
-                        Log(LogSeverityType.Error, functionErrorMessage + "Failed to parse goal declaration");
+                        Logger?.LogError("Failed to parse the file '{filePathAndName}': Failed to parse goal declaration", fileNameAndPath);
                         return false;
                     }
                     if (declaration != null)
                         track.Declarations.Add(declaration);
                 }
 
-                Pilot pilot = new Pilot(pilotNumber, pilotIdentifier);
+                Pilot pilot = new(pilotNumber, pilotIdentifier);
                 track.Pilot = pilot;
             }
             catch (Exception ex)
             {
-                Log(LogSeverityType.Error, functionErrorMessage + $" {ex.Message}");
+                Logger?.LogError(ex, "Failed to parse the file '{filePathAndName}'", fileNameAndPath);
                 return false;
             }
-            Log(LogSeverityType.Info, $"Successfully parsed file '{fileNameAndPath}'");
+            Logger?.LogInformation("Successfully parsed file '{filePathAndName}'", fileNameAndPath);
             return true;
         }
 
         #endregion
 
         #region Private methods
-        private static bool ParseHeaders(string[] headerLines, out int pilotNumber, out DateTime date,
-            out bool markerDrop_HasAdditionalLatitudeDecimals,
-            out int markerDrop_StartOfAdditionalLatitudeDecimals,
-            out int markerDrop_EndOfAdditionalLatitudeDecimals,
-            out bool markerDrop_HasAdditionalLongitudeDecimals,
-            out int markerDrop_StartOfAdditionalLongitudeDecimals,
-            out int markerDrop_EndOfAdditionalLongitudeDecimals, out bool declaration_HasAdditionalLatitudeDecimals,
-            out int declaration_StartOfAdditionalLatitudeDecimals,
-            out int declaration_EndOfAdditionalLatitudeDecimals,
-            out bool declaration_HasAdditionalLongitudeDecimals,
-            out int declaration_StartOfAdditionalLongitudeDecimals,
-            out int declaration_EndOfAdditionalLongitudeDecimals)
+        private static bool ParseHeaders(string[] headerLines, out int pilotNumber, out DateTime date)
         {
             date = DateTime.MinValue;
             pilotNumber = -1;
-            string functionErrorMessage = "Failed to parse header lines: ";
-            markerDrop_HasAdditionalLatitudeDecimals = false;
-            markerDrop_StartOfAdditionalLatitudeDecimals = -1;
-            markerDrop_EndOfAdditionalLatitudeDecimals = -1;
-            markerDrop_HasAdditionalLongitudeDecimals = false;
-            markerDrop_StartOfAdditionalLongitudeDecimals = -1;
-            markerDrop_EndOfAdditionalLongitudeDecimals = -1;
-            declaration_HasAdditionalLatitudeDecimals = false;
-            declaration_StartOfAdditionalLatitudeDecimals = -1;
-            declaration_EndOfAdditionalLatitudeDecimals = -1;
-            declaration_HasAdditionalLongitudeDecimals = false;
-            declaration_StartOfAdditionalLongitudeDecimals = -1;
-            declaration_EndOfAdditionalLongitudeDecimals = -1;
-
             foreach (string headerLine in headerLines)
             {
                 if (headerLine.StartsWith("HFDTE"))
                 {
                     string line = headerLine.Replace("HFDTE", "");
-                    int day;
-                    if (!int.TryParse(line[0..2], out day))
+                    if (!int.TryParse(line[0..2], out int day))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse day portion of date '{line[0..2]}' in '{line}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse day portion of date '{line[0..2]}' in '{line}'");
+                        Logger?.LogError("Failed to parse header lines: Could not parse day portion of date '{portion}' in '{line}'", line[0..2], line);
                         return false;
                     }
-                    int month;
-                    if (!int.TryParse(line[2..4], out month))
+                    if (!int.TryParse(line[2..4], out int month))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse month portion of date '{line[2..4]}' in '{line}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse month portion of date '{line[2..4]}' in '{line}'");
+                        Logger?.LogError("Failed to parse header lines: Could not parse month portion of date '{portion}' in '{line}'", line[2..4], line);
                         return false;
                     }
-                    int year;
-                    if (!int.TryParse(line[4..6], out year))
+                    if (!int.TryParse(line[4..6], out int year))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse year portion of date '{line[4..6]}' in '{line}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse year portion of date '{line[4..6]}' in '{line}'");
+                        Logger?.LogError("Failed to parse header lines: Could not parse year portion of date '{portion}' in '{line}'", line[4..6], line);
                         return false;
                     }
                     year += 2000;
@@ -376,93 +306,86 @@ namespace Coordinates.Parsers
                     string pilotNumberText = headerLine.Replace("HFPID", "");
                     if (!int.TryParse(pilotNumberText, out pilotNumber))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse the pilot number '{pilotNumberText}' in '{headerLine}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse the pilot number '{pilotNumberText}' in '{headerLine}'");
+                        Logger?.LogError("Failed to parse header lines: Could not parse pilot number '{pilotNumberText}' in '{headerLine}'", pilotNumberText, headerLine);
                         return false;
                     }
                 }
                 if (headerLine.StartsWith("HFXII:XX0:"))
                 {
-                    int numberOfAdditions;
-                    if (!int.TryParse(headerLine[10..12], out numberOfAdditions))
+                    if (!int.TryParse(headerLine[10..12], out int numberOfAdditions))
                     {
-                        Log(LogSeverityType.Error, functionErrorMessage + " Failed to parse number of additions from I-record");
+                        Logger?.LogError("Failed to parse header lines: Could not parse number of additions from I-record from portion '{portion}' in '{headerLine}'", headerLine[10..12], headerLine);
                         return false;
                     }
                     int offset = 12;
                     for (int index = 0; index < numberOfAdditions; index++)
                     {
                         offset = index * 7 + 12;
-                        int startPosition;
-                        if (!int.TryParse(headerLine[offset..(offset + 2)], out startPosition))
+                        if (!int.TryParse(headerLine[offset..(offset + 2)], out int startPosition))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse start position of addition no {index + 1}");
+                            Logger?.LogError("Failed to parse header lines: Could not parse start position of I-record addition no '{additionNumber}' from portion '{portion}' in '{headerLine}'", index + 1, headerLine[offset..(offset + 2)], headerLine);
                             return false;
                         }
                         offset += 2;
-                        int stopPosition;
-                        if (!int.TryParse(headerLine[(offset)..(offset + 2)], out stopPosition))
+                        if (!int.TryParse(headerLine[(offset)..(offset + 2)], out int stopPosition))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse stop position of addition no {index + 1}");
+                            Logger?.LogError("Failed to parse header lines: Could not parse stop position of I-record addition no '{additionNumber}' in from portion '{portion}' '{headerLine}'", index + 1, headerLine[(offset)..(offset + 2)], headerLine);
                             return false;
                         }
                         offset += 2;
                         string additionIdentifier = headerLine[(offset)..(offset + 3)];
 
-                        if (additionIdentifier.ToUpper() == "LAD")
+                        if (additionIdentifier.Equals("LAD", StringComparison.CurrentCultureIgnoreCase))
                         {
-                            markerDrop_HasAdditionalLatitudeDecimals = true;
-                            markerDrop_StartOfAdditionalLatitudeDecimals = startPosition - 1;//adjust to zero based index
-                            markerDrop_EndOfAdditionalLatitudeDecimals = stopPosition;
+                            MarkerDrop_HasAdditionalLatitudeDecimals = true;
+                            MarkerDrop_StartOfAdditionalLatitudeDecimals = startPosition - 1;//adjust to zero based index
+                            MarkerDrop_EndOfAdditionalLatitudeDecimals = stopPosition;
                         }
-                        if (additionIdentifier.ToUpper() == "LOD")
+                        if (additionIdentifier.Equals("LOD", StringComparison.CurrentCultureIgnoreCase))
                         {
-                            markerDrop_HasAdditionalLongitudeDecimals = true;
-                            markerDrop_StartOfAdditionalLongitudeDecimals = startPosition - 1;//adjust to zero based index
-                            markerDrop_EndOfAdditionalLongitudeDecimals = stopPosition;
+                            MarkerDrop_HasAdditionalLongitudeDecimals = true;
+                            MarkerDrop_StartOfAdditionalLongitudeDecimals = startPosition - 1;//adjust to zero based index
+                            MarkerDrop_EndOfAdditionalLongitudeDecimals = stopPosition;
                         }
                     }
 
                 }
                 if (headerLine.StartsWith("HFXII:XL1:"))
                 {
-                    int numberOfAdditions;
-                    if (!int.TryParse(headerLine[10..12], out numberOfAdditions))
+                    if (!int.TryParse(headerLine[10..12], out int numberOfAdditions))
                     {
-                        Log(LogSeverityType.Error, functionErrorMessage + " Failed to parse number of additions from I-record");
+                        Logger?.LogError("Failed to parse header lines: Could not parse number of additions from I-record from portion '{portion}' in '{headerLine}'", headerLine[10..12], headerLine);
                         return false;
                     }
-                    int offset = 12;
+
                     for (int index = 0; index < numberOfAdditions; index++)
                     {
-                        offset = index * 7 + 12;
-                        int startPosition;
-                        if (!int.TryParse(headerLine[offset..(offset + 2)], out startPosition))
+                        int offset = index * 7 + 12;
+                        if (!int.TryParse(headerLine[offset..(offset + 2)], out int startPosition))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse start position of addition no {index + 1}");
+                            Logger?.LogError("Failed to parse header lines: Could not parse start position of I-record addition no '{additionNumber}' from portion '{portion}' in '{headerLine}'", index + 1, headerLine[offset..(offset + 2)], headerLine);
                             return false;
                         }
                         offset += 2;
-                        int stopPosition;
-                        if (!int.TryParse(headerLine[(offset)..(offset + 2)], out stopPosition))
+                        if (!int.TryParse(headerLine[(offset)..(offset + 2)], out int stopPosition))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse stop position of addition no {index + 1}");
+                            Logger?.LogError("Failed to parse header lines: Could not parse stop position of I-record addition no '{additionNumber}' in from portion '{portion}' '{headerLine}'", index + 1, headerLine[(offset)..(offset + 2)], headerLine);
                             return false;
                         }
                         offset += 2;
                         string additionIdentifier = headerLine[(offset)..(offset + 3)];
 
-                        if (additionIdentifier.ToUpper() == "LAD")
+                        if (additionIdentifier.Equals("LAD", StringComparison.CurrentCultureIgnoreCase))
                         {
-                            declaration_HasAdditionalLatitudeDecimals = true;
-                            declaration_StartOfAdditionalLatitudeDecimals = startPosition - 1;//adjust to zero based index
-                            declaration_EndOfAdditionalLatitudeDecimals = stopPosition;
+                            Declaration_HasAdditionalLatitudeDecimals = true;
+                            Declaration_StartOfAdditionalLatitudeDecimals = startPosition - 1;//adjust to zero based index
+                            Declaration_EndOfAdditionalLatitudeDecimals = stopPosition;
                         }
-                        if (additionIdentifier.ToUpper() == "LOD")
+                        if (additionIdentifier.Equals("LOD", StringComparison.CurrentCultureIgnoreCase))
                         {
-                            declaration_HasAdditionalLongitudeDecimals = true;
-                            declaration_StartOfAdditionalLongitudeDecimals = startPosition - 1;//adjust to zero based index
-                            declaration_EndOfAdditionalLongitudeDecimals = stopPosition;
+                            Declaration_HasAdditionalLongitudeDecimals = true;
+                            Declaration_StartOfAdditionalLongitudeDecimals = startPosition - 1;//adjust to zero based index
+                            Declaration_EndOfAdditionalLongitudeDecimals = stopPosition;
                         }
                     }
                 }
@@ -482,32 +405,22 @@ namespace Coordinates.Parsers
         /// <param name="date">the date from the header</param>
         /// <param name="coordinate">output parameter. a trackpoint as coordinate object</param>
         /// <returns>true:success; false:error</returns>
-        private static bool ParseTrackPoint(string line, DateTime date,
-            bool trackPoint_HasAdditionalLatitudeDecimals,
-            int trackPoint_StartOfAdditionalLatitudeDecimals,
-            int trackPoint_EndOfAdditionalLatitudeDecimals,
-            bool trackPoint_HasAdditionalLongitudeDecimals,
-            int trackPoint_StartOfAdditionalLongitudeDecimals,
-            int trackPoint_EndOfAdditionalLongitudeDecimals,
-            out Coordinate coordinate)
+        private static bool ParseTrackPoint(string line, DateTime date, out Coordinate coordinate)
         {
-            string functionErrorMessage = $"Failed to parse track point:";
             coordinate = null;
 
-            DateTime timeStamp;
-            if (!ParseTimeStamp(line, date, out timeStamp))
+            if (!ParseTimeStamp(line, date, out DateTime timeStamp))
             {
-                //Debug.WriteLine(functionErrorMessage);
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse track point: Cannot parse timestamp from '{line}'", line);
                 return false;
             }
 
             double latitude;
-            if (trackPoint_HasAdditionalLatitudeDecimals)
+            if (TrackPoint_HasAdditionalLatitudeDecimals)
             {
-                if (!ParseLatitudeWithAdditionalDecimals(line[7..15], line[trackPoint_StartOfAdditionalLatitudeDecimals..trackPoint_EndOfAdditionalLatitudeDecimals], out latitude))
+                if (!ParseLatitudeWithAdditionalDecimals(line[7..15], line[TrackPoint_StartOfAdditionalLatitudeDecimals..TrackPoint_EndOfAdditionalLatitudeDecimals], out latitude))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage);
+                    Logger?.LogError("Failed to parse track point: Cannot parse latitude from portion '{portion}' and additional decimals '{additionDecimals}' in '{line}'", line[7..15], line[TrackPoint_StartOfAdditionalLatitudeDecimals..TrackPoint_EndOfAdditionalLatitudeDecimals], line);
                     return false;
                 }
             }
@@ -515,17 +428,16 @@ namespace Coordinates.Parsers
             {
                 if (!ParseLatitude(line[7..15], out latitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage);
-                    Log(LogSeverityType.Error, functionErrorMessage);
+                    Logger?.LogError("Failed to parse track point: Cannot parse latitude from portion '{portion}' in '{line}'", line[7..15], line);
                     return false;
                 }
             }
             double longitude;
-            if (trackPoint_HasAdditionalLongitudeDecimals)
+            if (TrackPoint_HasAdditionalLongitudeDecimals)
             {
-                if (!ParseLongitudeWithAdditionalDecimals(line[15..24], line[trackPoint_StartOfAdditionalLongitudeDecimals..trackPoint_EndOfAdditionalLongitudeDecimals], out longitude))
+                if (!ParseLongitudeWithAdditionalDecimals(line[15..24], line[TrackPoint_StartOfAdditionalLongitudeDecimals..TrackPoint_EndOfAdditionalLongitudeDecimals], out longitude))
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage);
+                    Logger?.LogError("Failed to parse track point: Cannot parse longitude from portion '{portion}' and additional decimals '{additionDecimals}' in '{line}'", line[15..24], line[TrackPoint_StartOfAdditionalLongitudeDecimals..TrackPoint_EndOfAdditionalLongitudeDecimals], line);
                     return false;
                 }
             }
@@ -533,8 +445,7 @@ namespace Coordinates.Parsers
             {
                 if (!ParseLongitude(line[15..24], out longitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage);
-                    Log(LogSeverityType.Error, functionErrorMessage);
+                    Logger?.LogError("Failed to parse track point: Cannot parse longitude from portion '{portion}' in '{line}'", line[15..24], line);
                     return false;
                 }
             }
@@ -542,19 +453,15 @@ namespace Coordinates.Parsers
 
 
 
-            double altitudeBarometric;
-            if (!double.TryParse(line[25..30], out altitudeBarometric))
+            if (!double.TryParse(line[25..30], out double altitudeBarometric))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude '{line[25..30]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse barometric altitude '{line[25..30]}' in '{line}'");
+                Logger?.LogError("Failed to parse track point: Cannot parse barometric altitude from portion '{portion}' in '{line}'", line[25..30], line);
                 return false;
             }
 
-            double altitudeGPS;
-            if (!double.TryParse(line[30..35], out altitudeGPS))
+            if (!double.TryParse(line[30..35], out double altitudeGPS))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude '{line[30..35]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse GPS altitude '{line[30..35]}' in '{line}'");
+                Logger?.LogError("Failed to parse track point: Cannot parse GPS altitude from portion '{portion}' in '{line}'", line[30..35], line);
                 return false;
             }
             coordinate = new Coordinate(latitude, longitude, altitudeGPS, altitudeBarometric, timeStamp);
@@ -575,41 +482,27 @@ namespace Coordinates.Parsers
         /// <param name="referenceCoordinate">a reference coordinate to fill up the missing info from utm goal declaration. If the reference is null, the position of declaration will be used instead</param>
         /// <param name="declaration">output parameter. the declaration</param>
         /// <returns>true:success; false:error</returns>
-        private static bool ParseGoalDeclaration(string line, DateTime date, bool declaredAltitudeIsInFeet,double defaultGoalAltitude,
-            Coordinate referenceCoordinate,
-            bool declaration_HasAdditionalLatitudeDecimals,
-            int declaration_StartOfAdditionalLatitudeDecimals,
-            int declaration_EndOfAdditionalLatitudeDecimals,
-            bool declaration_HasAdditionalLongitudeDecimals,
-            int declaration_StartOfAdditionalLongitudeDecimals,
-            int declaration_EndOfAdditionalLongitudeDecimals,
-            out Declaration declaration)
+        private static bool ParseGoalDeclaration(string line, DateTime date, bool declaredAltitudeIsInFeet, Coordinate referenceCoordinate, out Declaration declaration)
         {
-            string functionErrorMessage = $"Failed to parse goal declaration:";
             declaration = null;
 
-            DateTime timeStamp;
-            if (!ParseTimeStamp(line, date, out timeStamp))
+            if (!ParseTimeStamp(line, date, out DateTime timeStamp))
             {
-                //Debug.WriteLine(functionErrorMessage);
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse goal declaration: Cannot parse timestamp from '{line}'", line);
                 return false;
             }
-            int goalNumber;
-            if (!int.TryParse(line[10..12], out goalNumber))
+            if (!int.TryParse(line[10..12], out int goalNumber))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal number '{line[10..12]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal number '{line[10..12]}' in '{line}'");
+                Logger?.LogError("Failed to parse goal declaration: Cannot parse goal number from portion '{portion}' in '{line}'", line[10..12], line);
                 return false;
             }
 
             double declarationLatitude;
-            if (declaration_HasAdditionalLatitudeDecimals)
+            if (Declaration_HasAdditionalLatitudeDecimals)
             {
-                if (!ParseLatitudeWithAdditionalDecimals(line[12..20], line[declaration_StartOfAdditionalLatitudeDecimals..declaration_EndOfAdditionalLatitudeDecimals], out declarationLatitude))
+                if (!ParseLatitudeWithAdditionalDecimals(line[12..20], line[Declaration_StartOfAdditionalLatitudeDecimals..Declaration_EndOfAdditionalLatitudeDecimals], out declarationLatitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"Failed to parse latitude at declaration position '{parts[2][0..8]}' in '{line}'");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse latitude at declaration position in '{line}'");
+                    Logger?.LogError("Failed to parse goal declaration: Cannot parse latitude from portion '{portion}' and additional decimals '{additionDecimals}' in '{line}'", line[12..20], line[Declaration_StartOfAdditionalLatitudeDecimals..Declaration_EndOfAdditionalLatitudeDecimals], line);
                     return false;
                 }
             }
@@ -617,18 +510,16 @@ namespace Coordinates.Parsers
             {
                 if (!ParseLatitude(line[12..20], out declarationLatitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"Failed to parse latitude at declaration position '{parts[2][0..8]}' in '{line}'");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse latitude at declaration position  in '{line}'");
+                    Logger?.LogError("Failed to parse goal declaration: Cannot parse latitude from portion '{portion}' in '{line}'", line[12..20], line);
                     return false;
                 }
             }
             double declarationLongitude;
-            if (declaration_HasAdditionalLongitudeDecimals)
+            if (Declaration_HasAdditionalLongitudeDecimals)
             {
-                if (!ParseLongitudeWithAdditionalDecimals(line[20..29], line[declaration_StartOfAdditionalLongitudeDecimals..declaration_EndOfAdditionalLongitudeDecimals], out declarationLongitude))
+                if (!ParseLongitudeWithAdditionalDecimals(line[20..29], line[Declaration_StartOfAdditionalLongitudeDecimals..Declaration_EndOfAdditionalLongitudeDecimals], out declarationLongitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"Failed to parse longitude at declaration position '{parts[2][8..17]}' in '{line}'");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse longitude at declaration position  in '{line}'");
+                    Logger?.LogError("Failed to parse goal declaration: Cannot parse longitude from portion '{portion}' and additional decimals '{additionDecimals}' in '{line}'", line[20..29], line[Declaration_StartOfAdditionalLongitudeDecimals..Declaration_EndOfAdditionalLongitudeDecimals], line);
                     return false;
                 }
             }
@@ -636,25 +527,20 @@ namespace Coordinates.Parsers
             {
                 if (!ParseLongitude(line[20..29], out declarationLongitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"Failed to parse longitude at declaration position '{parts[2][8..17]}' in '{line}'");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse longitude at declaration position  in '{line}'");
+                    Logger?.LogError("Failed to parse goal declaration: Cannot parse longitude from portion '{portion}' in '{line}'", line[20..29], line);
                     return false;
                 }
             }
 
-            double declarationPositonAltitudeBarometric;
-            if (!double.TryParse(line[30..35], out declarationPositonAltitudeBarometric))
+            if (!double.TryParse(line[30..35], out double declarationPositonAltitudeBarometric))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude at declaration position '{parts[2][18..23]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse barometric altitude at declaration position in '{line}'");
+                Logger?.LogError("Failed to parse goal declaration: Cannot parse barometric altitude from portion '{portion}' in '{line}'", line[30..35], line);
                 return false;
             }
 
-            double declarationPositionAltitudeGPS;
-            if (!double.TryParse(line[35..40], out declarationPositionAltitudeGPS))
+            if (!double.TryParse(line[35..40], out double declarationPositionAltitudeGPS))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse GPS altitude at declaration position '{parts[2][23..28]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse GPS altitude at declaration position in '{line}'");
+                Logger?.LogError("Failed to parse goal declaration: Cannot parse GPS altitude from portion '{portion}' in '{line}'", line[35..40], line);
                 return false;
             }
             string declarationText = line[43..^0];
@@ -663,10 +549,10 @@ namespace Coordinates.Parsers
 
                 string[] parts = declarationText.Split(',');
                 string[] locations = parts[0].Split('/');
-                int eastingDigits = -1;
-                int eastingUTM = -1;
-                int northingDigits = -1;
-                int northingUTM = -1;
+                int eastingDigits;
+                int eastingUTM;
+                int northingDigits;
+                int northingUTM;
                 if (locations.Length == 2)
                 {
                     if (!string.IsNullOrWhiteSpace(locations[0]))
@@ -674,13 +560,13 @@ namespace Coordinates.Parsers
                         eastingDigits = locations[0].Length;
                         if (!int.TryParse(locations[0], out eastingUTM))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse easting of declared goal in '{declarationText}'");
+                            Logger?.LogError("Failed to parse goal declaration: Cannot parse easting from portion '{portion}' in '{line}'", locations[0], line);
                             return false;
                         }
                     }
                     else
                     {
-                        Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse easting of declared goal in '{declarationText}'");
+                        Logger?.LogError("Failed to parse goal declaration: Cannot parse easting from portion '{portion}' in '{line}'", locations[0], line);
                         return false;
                     }
                     if (!string.IsNullOrWhiteSpace(locations[1]))
@@ -688,20 +574,20 @@ namespace Coordinates.Parsers
                         northingDigits = locations[1].Length;
                         if (!int.TryParse(locations[1], out northingUTM))
                         {
-                            Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse northing of declared goal in '{declarationText}'");
+                            Logger?.LogError("Failed to parse goal declaration: Cannot parse northing from portion '{portion}' in '{line}'", locations[1], line);
                             return false;
                         }
                     }
                     else
                     {
-                        Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse northing of declared goal in '{declarationText}'");
+                        Logger?.LogError("Failed to parse goal declaration: Cannot parse northing from portion '{portion}' in '{line}'", locations[1], line);
                         return false;
                     }
 
                 }
                 else
                 {
-                    Log(LogSeverityType.Error, functionErrorMessage + $" Failed to parse easting and northing of declared goal in '{declarationText}'");
+                    Logger?.LogError("Failed to parse goal declaration: Cannot parse easting and northing from portion '{portion}' in '{line}'", declarationText, line);
                     return false;
                 }
 
@@ -712,11 +598,9 @@ namespace Coordinates.Parsers
                     if (!string.IsNullOrWhiteSpace(parts[1]))
                     {
                         string altitudePart = parts[1].Replace("ft", "").Replace("m", "");
-                        int declaredAltitude;
-                        if (!int.TryParse(altitudePart, out declaredAltitude))
+                        if (!int.TryParse(altitudePart, out int declaredAltitude))
                         {
-                            //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration altitude portion '{parts[1]}' in '{line}'");
-                            Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal declaration altitude portion '{altitudePart}' in '{line}'");
+                            Logger?.LogError("Failed to parse goal declaration: Cannot parse altitude from portion '{portion}' in '{line}'", altitudePart, line);
                             return false;
                         }
                         if (declaredAltitudeIsInFeet)
@@ -727,15 +611,15 @@ namespace Coordinates.Parsers
                     else
                     {
                         hasPilotDelaredGoalAltitude = false;
-                        declaredAltitudeInMeter = defaultGoalAltitude;
-                        Log(LogSeverityType.Warning, $"No altitude declared for Goal No. '{goalNumber}'. Default goal altitude of '{defaultGoalAltitude}' will be assumed");
+                        declaredAltitudeInMeter = 0.0;
+                        Logger?.LogWarning("No altitude declared for Goal No. '{goalNumber}'. Altitude of 0 will be assumed", goalNumber);
                     }
                 }
                 else
                 {
                     hasPilotDelaredGoalAltitude = false;
-                    declaredAltitudeInMeter = defaultGoalAltitude;
-                    Log(LogSeverityType.Warning, $"No altitude declared for Goal No. '{goalNumber}'. Default goal altitude of '{defaultGoalAltitude}' will be assumed");
+                    declaredAltitudeInMeter = 0.0;
+                    Logger?.LogWarning("No altitude declared for Goal No. '{goalNumber}'. Altitude of 0 will be assumed", goalNumber);
                 }
 
                 CoordinateSharp.Coordinate coordinateSharp;
@@ -756,7 +640,7 @@ namespace Coordinates.Parsers
                     }
                     if (northingDigits == 6)
                     {
-                        northingUTM += (int)(Math.Floor(coordinateSharp.UTM.Northing / Math.Pow(10, northingDigits)) * Math.Pow(10, northingDigits));
+                        goalNorthingUTM += (int)(Math.Floor(coordinateSharp.UTM.Northing / Math.Pow(10, northingDigits)) * Math.Pow(10, northingDigits));
                     }
 
                     if (eastingDigits != 6)
@@ -766,7 +650,7 @@ namespace Coordinates.Parsers
 
                     }
 
-                    CoordinateSharp.UniversalTransverseMercator utm = new CoordinateSharp.UniversalTransverseMercator(utmGridZone, goalEastingUTM, goalNorthingUTM);
+                    CoordinateSharp.UniversalTransverseMercator utm = new(utmGridZone, goalEastingUTM, goalNorthingUTM);
 
                     CoordinateSharp.Coordinate coordinate = CoordinateSharp.UniversalTransverseMercator.ConvertUTMtoLatLong(utm);
 
@@ -790,7 +674,7 @@ namespace Coordinates.Parsers
                     }
                     if (northingDigits == 6)
                     {
-                        northingUTM += (int)(Math.Floor(coordinateSharp.UTM.Northing / Math.Pow(10, northingDigits)) * Math.Pow(10, northingDigits));
+                        goalNorthingUTM += (int)(Math.Floor(coordinateSharp.UTM.Northing / Math.Pow(10, northingDigits)) * Math.Pow(10, northingDigits));
                     }
 
                     if (eastingDigits != 6)
@@ -800,7 +684,7 @@ namespace Coordinates.Parsers
 
                     }
 
-                    CoordinateSharp.UniversalTransverseMercator utm = new CoordinateSharp.UniversalTransverseMercator(utmGridZone, goalEastingUTM, goalNorthingUTM);
+                    CoordinateSharp.UniversalTransverseMercator utm = new(utmGridZone, goalEastingUTM, goalNorthingUTM);
 
                     CoordinateSharp.Coordinate coordinate = CoordinateSharp.UniversalTransverseMercator.ConvertUTMtoLatLong(utm);
 
@@ -809,16 +693,15 @@ namespace Coordinates.Parsers
                     double distance = CoordinateHelpers.Calculate2DDistanceHavercos(declaredGoal, positionAtDeclaration);
                     if (distance > 70e3)
                     {
-
-                        Log(LogSeverityType.Warning, $"Suspicious declaration of gaol {goalNumber}: {locations[0]}/{locations[1]}");
+                        Logger?.LogWarning("Suspicious declaration of gaol {goalNumber}: {locations[0]}/{locations[1]}", goalNumber, locations[0], locations[1]);
                     }
                 }
 
-                declaration = new Declaration(goalNumber, declaredGoal, positionAtDeclaration);
+                declaration = new Declaration(goalNumber, declaredGoal, positionAtDeclaration, hasPilotDelaredGoalAltitude, eastingUTM, northingUTM);
             }
             else
             {
-                Log(LogSeverityType.Warning, "Declaration contained no coordinates and therefore has not been added to the track");
+                Logger?.LogWarning("Declaration contained no coordinates and therefore has not been added to the track");
             }
 
             return true;
@@ -834,40 +717,27 @@ namespace Coordinates.Parsers
         /// <param name="date">the data form the header</param>
         /// <param name="markerDrop">output parameter. the marker drop as object</param>
         /// <returns>true:success; false: error</returns>
-        private static bool ParseMarkerDrop(string line, DateTime date,
-            bool markerDrop_HasAdditionalLatitudeDecimals,
-            int markerDrop_StartOfAdditionalLatitudeDecimals,
-            int markerDrop_EndOfAdditionalLatitudeDecimals,
-            bool markerDrop_HasAdditionalLongitudeDecimals,
-            int markerDrop_StartOfAdditionalLongitudeDecimals,
-            int markerDrop_EndOfAdditionalLongitudeDecimals,
-            out MarkerDrop markerDrop)
+        private static bool ParseMarkerDrop(string line, DateTime date, out MarkerDrop markerDrop)
         {
-            string functionErrorMessage = $"Failed to parse marker drop:";
             markerDrop = null;
 
-            DateTime timeStamp;
-            if (!ParseTimeStamp(line, date, out timeStamp))
+            if (!ParseTimeStamp(line, date, out DateTime timeStamp))
             {
-                //Debug.WriteLine(functionErrorMessage);
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse marker drop: Cannot parse timestamp from '{line}'", line);
                 return false;
             }
-            int markerNumber;
-            if (!int.TryParse(line[10..12], out markerNumber))
+            if (!int.TryParse(line[10..12], out int markerNumber))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse marker number '{line[10..12]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse marker number '{line[10..12]}' in '{line}'");
+                Logger?.LogError("Failed to parse marker drop: Cannot parse marker number from portion '{portion}' in '{line}'", line[10..12], line);
                 return false;
             }
 
             double latitude;
-            if (markerDrop_HasAdditionalLatitudeDecimals)
+            if (MarkerDrop_HasAdditionalLatitudeDecimals)
             {
-                if (!ParseLatitudeWithAdditionalDecimals(line[12..20], line[markerDrop_StartOfAdditionalLatitudeDecimals..markerDrop_EndOfAdditionalLatitudeDecimals], out latitude))
+                if (!ParseLatitudeWithAdditionalDecimals(line[12..20], line[MarkerDrop_StartOfAdditionalLatitudeDecimals..MarkerDrop_EndOfAdditionalLatitudeDecimals], out latitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"Failed to parse marker drop latitude '{line[12..20]}' in '{line}'");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse marker drop latitude '{line[12..20]}' in '{line}'");
+                    Logger?.LogError("Failed to parse marker drop: Cannot parse latitude from portion '{portion}' and additional decimals '{additionDecimals}' in '{line}'", line[12..20], line[MarkerDrop_StartOfAdditionalLatitudeDecimals..MarkerDrop_EndOfAdditionalLatitudeDecimals], line);
                     return false;
                 }
             }
@@ -875,18 +745,16 @@ namespace Coordinates.Parsers
             {
                 if (!ParseLatitude(line[12..20], out latitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"Failed to parse marker drop latitude '{line[12..20]}' in '{line}'");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse marker drop latitude '{line[12..20]}' in '{line}'");
+                    Logger?.LogError("Failed to parse marker drop: Cannot parse latitude from portion '{portion}' in '{line}'", line[12..20], line);
                     return false;
                 }
             }
             double longitude;
-            if (markerDrop_HasAdditionalLongitudeDecimals)
+            if (MarkerDrop_HasAdditionalLongitudeDecimals)
             {
-                if (!ParseLongitudeWithAdditionalDecimals(line[20..29], line[markerDrop_StartOfAdditionalLongitudeDecimals..markerDrop_EndOfAdditionalLongitudeDecimals], out longitude))
+                if (!ParseLongitudeWithAdditionalDecimals(line[20..29], line[MarkerDrop_StartOfAdditionalLongitudeDecimals..MarkerDrop_EndOfAdditionalLongitudeDecimals], out longitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"Failed to parse marker drop longitude '{line[20..29]}' in '{line}'");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse marker drop longitude '{line[20..29]}' in '{line}'");
+                    Logger?.LogError("Failed to parse marker drop: Cannot parse longitude from portion '{portion}' and additional decimals '{additionDecimals}' in '{line}'", line[20..29], line[MarkerDrop_StartOfAdditionalLongitudeDecimals..MarkerDrop_EndOfAdditionalLongitudeDecimals], line);
                     return false;
                 }
             }
@@ -894,28 +762,23 @@ namespace Coordinates.Parsers
             {
                 if (!ParseLongitude(line[20..29], out longitude))
                 {
-                    //Debug.WriteLine(functionErrorMessage + $"Failed to parse marker drop longitude '{line[20..29]}' in '{line}'");
-                    Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse marker drop longitude '{line[20..29]}' in '{line}'");
+                    Logger?.LogError("Failed to parse marker drop: Cannot parse longitude from portion '{portion}' in '{line}'", line[20..29], line);
                     return false;
                 }
             }
 
-            double altitudeBarometric;
-            if (!double.TryParse(line[30..35], out altitudeBarometric))
+            if (!double.TryParse(line[30..35], out double altitudeBarometric))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude '{line[30..35]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse barometric altitude '{line[30..35]}' in '{line}'");
+                Logger?.LogError("Failed to parse marker drop: Cannot parse barometric altitude from portion '{portion}' in '{line}'", line[30..35], line);
                 return false;
             }
 
-            double altitudeGPS;
-            if (!double.TryParse(line[35..40], out altitudeGPS))
+            if (!double.TryParse(line[35..40], out double altitudeGPS))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude '{line[35..40]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse barometric altitude '{line[35..40]}' in '{line}'");
+                Logger?.LogError("Failed to parse marker drop: Cannot parse GPS altitude from portion '{portion}' in '{line}'", line[35..40], line);
                 return false;
             }
-            Coordinate coordinate = new Coordinate(latitude, longitude, altitudeGPS, altitudeBarometric, timeStamp);
+            Coordinate coordinate = new(latitude, longitude, altitudeGPS, altitudeBarometric, timeStamp);
             markerDrop = new MarkerDrop(markerNumber, coordinate);
             return true;
         }
@@ -936,23 +799,17 @@ namespace Coordinates.Parsers
                 factor = -1.0;
             else
             {
-                factor = double.NaN;
-                //Debug.WriteLine($"Failed to parse latitude text. Unexpected suffix '{latitudeText[^1]}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text. Unexpected suffix '{latitudeText[^1]}'");
+                Logger?.LogError("Failed to parse latitude text. Unexpected suffix '{suffix}'", latitudeText[^1]);
                 return false;
             }
-            double fullAngle = double.NaN;
-            if (!double.TryParse(latitudeText[0..2], out fullAngle))
+            if (!double.TryParse(latitudeText[0..2], out double fullAngle))
             {
-                //Debug.WriteLine($"Failed to parse latitude text '{latitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text '{latitudeText}'");
+                Logger?.LogError("Failed to parse latitude full angle from '{latitudeText}'", latitudeText[0..2]);
                 return false;
             }
-            double decimalAngle = double.NaN;
-            if (!double.TryParse(latitudeText[2..7], out decimalAngle))
+            if (!double.TryParse(latitudeText[2..7], out double decimalAngle))
             {
-                //Debug.WriteLine($"Failed to parse latitude text '{latitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text '{latitudeText}'");
+                Logger?.LogError("Failed to parse latitude decimal angle '{latitudeText}'", latitudeText[2..7]);
                 return false;
             }
             decimalAngle /= 60000.0;//divided by 1000 to get decimal value, divided by 60 to get from angle minutes to decimal angles
@@ -971,28 +828,21 @@ namespace Coordinates.Parsers
                 factor = -1.0;
             else
             {
-                factor = double.NaN;
-                //Debug.WriteLine($"Failed to parse latitude text. Unexpected suffix '{latitudeText[^1]}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text. Unexpected suffix '{standardLatitudeText[^1]}'");
+                Logger?.LogError("Failed to parse latitude text. Unexpected suffix '{suffix}'", standardLatitudeText[^1]);
                 return false;
             }
-            double fullAngle = double.NaN;
-            if (!double.TryParse(standardLatitudeText[0..2], out fullAngle))
+            if (!double.TryParse(standardLatitudeText[0..2], out double fullAngle))
             {
-                //Debug.WriteLine($"Failed to parse latitude text '{latitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text '{standardLatitudeText}'");
+                Logger?.LogError("Failed to parse latitude full angle from '{latitudeText}'", standardLatitudeText[0..2]);
                 return false;
             }
-            double decimalAngle = double.NaN;
-            if (!double.TryParse(standardLatitudeText[2..7] + additionalDecimals, out decimalAngle))
+            if (!double.TryParse(standardLatitudeText[2..7] + additionalDecimals, out double decimalAngle))
             {
-                //Debug.WriteLine($"Failed to parse latitude text '{latitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text '{standardLatitudeText}'");
+                Logger?.LogError("Failed to parse latitude decimal angle '{latitudeText}'", standardLatitudeText[2..7] + additionalDecimals);
                 return false;
             }
             if (decimalAngle > 0.0)
             {
-                //decimalAngle /= 60000.0;//divided by 1000 to get decimal value, divided by 60 to get from angle minutes to decimal angles
                 int divider = (standardLatitudeText[2..7] + additionalDecimals).Length - 2;//first two digits are integer the rest are decimal places
                 decimalAngle /= (Math.Pow(10.0, divider) * 60.0);//divide two get decimal places right and convert from minutes to degrees
             }
@@ -1016,23 +866,17 @@ namespace Coordinates.Parsers
                 factor = -1.0;
             else
             {
-                factor = double.NaN;
-                //Debug.WriteLine($"Failed to parse longitude text. Unexpected suffix '{longitudeText[^1]}'");
-                Log(LogSeverityType.Error, $"Failed to parse longitude text. Unexpected suffix '{longitudeText[^1]}'");
+                Logger?.LogError("Failed to parse longitude text. Unexpected suffix '{suffix}'", longitudeText[^1]);
                 return false;
             }
-            double fullAngle = double.NaN;
-            if (!double.TryParse(longitudeText[0..3], out fullAngle))
+            if (!double.TryParse(longitudeText[0..3], out double fullAngle))
             {
-                //Debug.WriteLine($"Failed to parse longitude text '{longitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse longitude text '{longitudeText}'");
+                Logger?.LogError("Failed to parse longitude full angle '{longitudeText}'", longitudeText[0..3]);
                 return false;
             }
-            double decimalAngle = double.NaN;
-            if (!double.TryParse(longitudeText[3..8], out decimalAngle))
+            if (!double.TryParse(longitudeText[3..8], out double decimalAngle))
             {
-                //Debug.WriteLine($"Failed to parse longitude text '{longitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse longitude text '{longitudeText}'");
+                Logger?.LogError("Failed to parse longitude decimal angle '{longitudeText}'", longitudeText[3..8]);
                 return false;
             }
             decimalAngle /= 60000.0;//divided by 1000 to get decimal value, divided by 60 to get from angle minutes to decimal degree
@@ -1051,26 +895,19 @@ namespace Coordinates.Parsers
                 factor = -1.0;
             else
             {
-                factor = double.NaN;
-                //Debug.WriteLine($"Failed to parse standardLongitude text. Unexpected suffix '{standardLongitudeText[^1]}'");
-                Log(LogSeverityType.Error, $"Failed to parse standardLongitude text. Unexpected suffix '{standardLongitudeText[^1]}'");
+                Logger?.LogError("Failed to parse longitude text. Unexpected suffix '{suffix}'", standardLongitudeText[^1]);
                 return false;
             }
-            double fullAngle = double.NaN;
-            if (!double.TryParse(standardLongitudeText[0..3], out fullAngle))
+            if (!double.TryParse(standardLongitudeText[0..3], out double fullAngle))
             {
-                //Debug.WriteLine($"Failed to parse standardLongitude text '{standardLongitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse standardLongitude text '{standardLongitudeText}'");
+                Logger?.LogError("Failed to parse longitude full angle '{longitudeText}'", standardLongitudeText[0..3]);
                 return false;
             }
-            double decimalAngle = double.NaN;
-            if (!double.TryParse(standardLongitudeText[3..8] + additinalDecimals, out decimalAngle))
+            if (!double.TryParse(standardLongitudeText[3..8] + additinalDecimals, out double decimalAngle))
             {
-                //Debug.WriteLine($"Failed to parse standardLongitude text '{standardLongitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse standardLongitude text '{standardLongitudeText}'");
+                Logger?.LogError("Failed to parse longitude decimal angle '{longitudeText}'", standardLongitudeText[3..8] + additinalDecimals);
                 return false;
             }
-            //decimalAngle /= 60000.0;//divided by 1000 to get decimal value, divided by 60 to get from angle minutes to decimal angles
             if (decimalAngle > 0.0)
             {
                 int divider = (standardLongitudeText[3..8] + additinalDecimals).Length - 2;//first two digits are integer the rest are decimal places
@@ -1089,51 +926,63 @@ namespace Coordinates.Parsers
         /// <returns>true:success: false:error</returns>
         private static bool ParseTimeStamp(string line, DateTime date, out DateTime timeStamp)
         {
-            string functionErrorMessage = "Failed to parse time:";
             string time = line[1..7];
-            int hours;
             timeStamp = date;
-            if (!int.TryParse(time[0..2], out hours))
+            if (!int.TryParse(time[0..2], out int hours))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse hour portion '{time[0..2]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse hour portion '{time[0..2]}' in '{line}'");
+                Logger?.LogError("Failed to parse time: Cannot parse hour portion '{portion}' in '{line}'", time[0..2], line);
                 return false;
             }
-            int minutes;
-            if (!int.TryParse(time[2..4], out minutes))
+            if (!int.TryParse(time[2..4], out int minutes))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse minute portion '{time[2..4]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse minute portion '{time[2..4]}' in '{line}'");
+                Logger?.LogError("Failed to parse time: Cannot parse minute portion '{portion}' in '{line}'", time[2..4], line);
                 return false;
             }
-            int seconds;
-            if (!int.TryParse(time[4..6], out seconds))
+            if (!int.TryParse(time[4..6], out int seconds))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse second portion '{time[4..6]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse second portion '{time[4..6]}' in '{line}'");
+                Logger?.LogError("Failed to parse time: Cannot parse second portion '{portion}' in '{line}'", time[4..6], line);
                 return false;
             }
             timeStamp = date.AddHours(hours).AddMinutes(minutes).AddSeconds(seconds);
             return true;
         }
 
-        private static void Log(LogSeverityType logSeverity, string text)
+        private static void ResetProperties()
         {
-            Logger.Log((object)"Balloon Live Parser", logSeverity, text);
-        }
+            MarkerDrop_HasAdditionalLatitudeDecimals = false;
+            MarkerDrop_StartOfAdditionalLatitudeDecimals = -1;
+            MarkerDrop_EndOfAdditionalLatitudeDecimals = -1;
 
+            MarkerDrop_HasAdditionalLongitudeDecimals = false;
+            MarkerDrop_StartOfAdditionalLongitudeDecimals = -1;
+            MarkerDrop_EndOfAdditionalLongitudeDecimals = -1;
+
+            Declaration_HasAdditionalLatitudeDecimals = false;
+            Declaration_StartOfAdditionalLatitudeDecimals = -1;
+            Declaration_EndOfAdditionalLatitudeDecimals = -1;
+
+            Declaration_HasAdditionalLongitudeDecimals = false;
+            Declaration_StartOfAdditionalLongitudeDecimals = -1;
+            Declaration_EndOfAdditionalLongitudeDecimals = -1;
+
+            TrackPoint_HasAdditionalLatitudeDecimals = false;
+            TrackPoint_StartOfAdditionalLatitudeDecimals = -1;
+            TrackPoint_EndOfAdditionalLatitudeDecimals = -1;
+
+            TrackPoint_HasAdditionalLongitudeDecimals = false;
+            TrackPoint_StartOfAdditionalLongitudeDecimals = -1;
+            TrackPoint_EndOfAdditionalLongitudeDecimals = -1;
+        }
 
         private static bool ParseSourceEvent(string line, DateTime date, out DateTime timeStamp, out bool isPrimarySource, out bool isBallonLiveSensor, out string blsSerialNumber)
         {
-            string functionErrorMessage = "Failed to parse position source event";
-            timeStamp = DateTime.MinValue;
             isPrimarySource = false;
             isBallonLiveSensor = false;
             blsSerialNumber = string.Empty;
 
             if (!ParseTimeStamp(line, date, out timeStamp))
             {
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse position source event: Cannot parse timestamp from '{line}'", line);
                 return false;
             }
             if (line.Contains("XS0"))
@@ -1146,7 +995,7 @@ namespace Coordinates.Parsers
             }
             else
             {
-                Log(LogSeverityType.Warning, $"Unknown source event '{line[7..10]}' in line '{line}'");
+                Logger?.LogWarning("Unknown source event '{sourceEvent}' in line '{line}'", line[7..10], line);
             }
 
             if (line.Contains("INT"))
@@ -1160,163 +1009,12 @@ namespace Coordinates.Parsers
             }
             else
             {
-                Log(LogSeverityType.Warning, $"Unknown position source '{line[10..13]}' in line '{line}'");
+                Logger?.LogWarning("Unknown position source '{sourceEvent}' in line '{line}'", line[10..13], line);
             }
 
             return true;
 
         }
-
-
-        ///// <summary>
-        ///// Parses a line with a goal declaration (EttttttXL1ddññññ*/ëëëë*,hhhh#,nnnnnnnNeeeeeeeeEbbbbbgggggaaassddd0000)
-        ///// <para>where t:timestamp d:number of declared goal ñ:goal northing in utm ë:goal easting in utm (*:the exact format in specified in the header at 'LXXX declaration digits') h:declared height (#: the unit is specified in the header)
-        ///// <para>n:northing e:easting b:barometric altitude g:gps altitude</para>
-        ///// <para>a:accuracy s:number of satellites d: engine noise level and rpm 0:carrier return and line feed</para>
-        ///// <para>e.g E105850XL101123456/987654,1500,4839658N00858176EA0000000537000224940000</para>
-        ///// </summary>
-        ///// <param name="line">the line in the file</param>
-        ///// <param name="date">the date form the header</param>
-        ///// <param name="declaredAltitudeIsInFeet">true: declared height is in feet; false: declared height is in meter</param>
-        ///// <param name="northingDigits">expected number of digits for goal northing</param>
-        ///// <param name="eastingDigits">expected number of digits for goal easting</param>
-        ///// <param name="referenceCoordinate">a reference coordinate to fill up the missing info from utm goal declaration. If the reference is null, the position of declaration will be used instead</param>
-        ///// <param name="declaredGoal">output parameter. the declared goal</param>
-        ///// <returns>true:success; false:error</returns>
-        //private static bool ParseGoalDeclaration(string line, DateTime date, bool declaredAltitudeIsInFeet, int northingDigits, int eastingDigits, Coordinate referenceCoordinate, out DeclaredGoal declaredGoal)
-        //{
-        //    string functionErrorMessage = $"Failed to parse goal declaration:";
-        //    declaredGoal = null;
-
-        //    DateTime timeStamp;
-        //    if (!ParseTimeStamp(line, date, out timeStamp))
-        //    {
-        //        //Debug.WriteLine(functionErrorMessage);
-        //        Log(LogSeverityType.Error, functionErrorMessage);
-        //        return false;
-        //    }
-        //    int goalNumber;
-        //    if (!int.TryParse(line[10..12], out goalNumber))
-        //    {
-        //        //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal number '{line[10..12]}' in '{line}'");
-        //        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal number '{line[10..12]}' in '{line}'");
-        //        return false;
-        //    }
-        //    int eastingUTM;
-        //    //int eastingStartCharacter = northingStartCharacter + northingDigits + 1;
-        //    int eastingStartCharacter = 12;
-        //    if (!int.TryParse(line[eastingStartCharacter..(eastingStartCharacter + eastingDigits)], out eastingUTM))
-        //    {
-        //        //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration easting portion '{line[eastingStartCharacter..(eastingStartCharacter + eastingDigits)]}' in '{line}'");
-        //        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal declaration easting portion '{line[eastingStartCharacter..(eastingStartCharacter + eastingDigits)]}' in '{line}'");
-        //        return false;
-        //    }
-        //    int northingUTM;
-        //    //int northingStartCharacter = 12;
-        //    int northingStartCharacter = eastingStartCharacter + eastingDigits + 1;
-        //    if (!int.TryParse(line[northingStartCharacter..(northingStartCharacter + northingDigits)], out northingUTM))
-        //    {
-        //        //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration northing portion '{line[northingStartCharacter..(northingStartCharacter + northingDigits)]}' in '{line}'");
-        //        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal declaration northing portion '{line[northingStartCharacter..(northingStartCharacter + northingDigits)]}' in '{line}'");
-        //        return false;
-        //    }
-        //    string[] parts = line.Split(',');
-        //    int declarationPartIndex = 2;
-        //    double declaredAltitudeInMeter;
-        //    if (parts.Length == 2)
-        //    {
-        //        if (!string.IsNullOrWhiteSpace(parts[1]))
-        //        {
-        //            string altitudePart = parts[1].Replace("ft", "").Replace("m", "");
-        //            int declaredAltitude;
-        //            if (!int.TryParse(altitudePart, out declaredAltitude))
-        //            {
-        //                //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration altitude portion '{parts[1]}' in '{line}'");
-        //                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal declaration altitude portion '{altitudePart}' in '{line}'");
-        //                return false;
-        //            }
-        //            if (declaredAltitudeIsInFeet)
-        //                declaredAltitudeInMeter = CoordinateHelpers.ConvertToMeter((double)declaredAltitude);
-        //            else
-        //                declaredAltitudeInMeter = (double)declaredAltitude;
-        //        }
-        //        else
-        //        {
-        //            declaredAltitudeInMeter = 0.0;
-        //            Log(LogSeverityType.Warning, $"No altitude declared for Goal No. '{goalNumber}'. Altitude of 0 will be assumed");
-        //        }
-        //    }
-        //    else
-        //    {
-        //        declaredAltitudeInMeter = 0.0;
-        //        Log(LogSeverityType.Warning, $"No altitude declared for Goal No. '{goalNumber}'. Altitude of 0 will be assumed");
-        //        declarationPartIndex = 1;
-        //    }
-        //    double declarationLatitude;
-        //    if (!ParseLatitude(parts[declarationPartIndex][0..8], out declarationLatitude))
-        //    {
-        //        //Debug.WriteLine(functionErrorMessage + $"Failed to parse latitude at declaration position '{parts[2][0..8]}' in '{line}'");
-        //        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse latitude at declaration position '{parts[2][0..8]}' in '{line}'");
-        //        return false;
-        //    }
-        //    double declarationLongitude;
-        //    if (!ParseLongitude(parts[declarationPartIndex][8..17], out declarationLongitude))
-        //    {
-        //        //Debug.WriteLine(functionErrorMessage + $"Failed to parse longitude at declaration position '{parts[2][8..17]}' in '{line}'");
-        //        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse longitude at declaration position '{parts[2][8..17]}' in '{line}'");
-        //        return false;
-        //    }
-
-        //    double declarationPositonAltitudeBarometric;
-        //    if (!double.TryParse(parts[declarationPartIndex][18..23], out declarationPositonAltitudeBarometric))
-        //    {
-        //        //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude at declaration position '{parts[2][18..23]}' in '{line}'");
-        //        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse barometric altitude at declaration position '{parts[2][18..23]}' in '{line}'");
-        //        return false;
-        //    }
-
-        //    double declarationPositionAltitudeGPS;
-        //    if (!double.TryParse(parts[declarationPartIndex][23..28], out declarationPositionAltitudeGPS))
-        //    {
-        //        //Debug.WriteLine(functionErrorMessage + $"Failed to parse GPS altitude at declaration position '{parts[2][23..28]}' in '{line}'");
-        //        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse GPS altitude at declaration position '{parts[2][23..28]}' in '{line}'");
-        //        return false;
-        //    }
-
-        //    CoordinateSharp.Coordinate coordinateSharp;
-        //    if (referenceCoordinate == null)
-        //        coordinateSharp = new CoordinateSharp.Coordinate(declarationLatitude, declarationLongitude);
-        //    else
-        //        coordinateSharp = new CoordinateSharp.Coordinate(referenceCoordinate.Latitude, referenceCoordinate.Longitude);
-
-        //    string utmGridZone = coordinateSharp.UTM.LatZone + coordinateSharp.UTM.LongZone;
-        //    if (northingDigits < 6)
-        //    {
-        //        northingUTM *= 10;
-        //        northingUTM += (int)(Math.Floor(coordinateSharp.UTM.Northing / Math.Pow(10, northingDigits + 1)) * Math.Pow(10, northingDigits + 1));
-        //    }
-        //    if (northingDigits == 6)
-        //    {
-        //        northingUTM += (int)(Math.Floor(coordinateSharp.UTM.Northing / Math.Pow(10, northingDigits)) * Math.Pow(10, northingDigits));
-        //    }
-
-        //    if (eastingDigits != 6)
-        //    {
-        //        eastingUTM *= 10;
-        //        eastingUTM += (int)(Math.Floor(coordinateSharp.UTM.Easting / Math.Pow(10, eastingDigits + 1)) * Math.Pow(10, eastingDigits + 1));
-        //    }
-
-        //    CoordinateSharp.UniversalTransverseMercator utm = new CoordinateSharp.UniversalTransverseMercator(utmGridZone, eastingUTM, northingUTM);
-
-        //    CoordinateSharp.Coordinate coordinate = CoordinateSharp.UniversalTransverseMercator.ConvertUTMtoLatLong(utm);
-
-        //    Coordinate goalDeclared = new Coordinate(coordinate.Latitude.DecimalDegree, coordinate.Longitude.DecimalDegree, declaredAltitudeInMeter, declaredAltitudeInMeter, timeStamp);
-        //    Coordinate positionAtDeclaration = new Coordinate(declarationLatitude, declarationLongitude, declarationPositionAltitudeGPS, declarationPositonAltitudeBarometric, timeStamp);
-
-        //    declaredGoal = new DeclaredGoal(goalNumber, goalDeclared, positionAtDeclaration);
-
-        //    return true;
-        //}
         #endregion
 
     }
