@@ -1,16 +1,16 @@
-﻿using LoggerComponent;
+﻿using LoggingConnector;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
-using System.Text;
-using System.Text.RegularExpressions;
 
 namespace Coordinates.Parsers
 {
     public static class FAILoggerParser
     {
+        private static readonly ILogger Logger = LogConnector.LoggerFactory.CreateLogger(nameof(FAILoggerParser));
+
         #region API
 
         /// <summary>
@@ -18,39 +18,36 @@ namespace Coordinates.Parsers
         /// </summary>
         /// <param name="fileNameAndPath">the file path and name of the .igc file</param>
         /// <param name="track">output parameter. the parsed track from the file</param>
-        /// <param name="defaultGoalAltitude">the default goal altitude in meters to be used when the pilot didn't declared one</param>
+        /// <param name="referenceCoordinate">provide a reference coordinate to be used complete the missing information of a goal declaration
+        /// <para>a goal declaration not in the zone 6/7 format is ambiguous</para>
+        /// <para>this is an optional parameter, the parse will use either marker drop 1 or the position at declaration if not reference point has been provided</para></param>
         /// <returns>true:success; false:error</returns>
-        public static bool ParseFile(string fileNameAndPath, out Track track, double defaultGoalAltitude = 0.0)
+        public static bool ParseFile(string fileNameAndPath, out Track track, Coordinate referenceCoordinate = null)
         {
-            //TODO make method async?
-
-            string functionErrorMessage = $"Failed to parse file '{fileNameAndPath}':";
             track = null;
 
-            FileInfo fileInfo = new FileInfo($@"{fileNameAndPath}");
+            FileInfo fileInfo = new($@"{fileNameAndPath}");
             if (!fileInfo.Exists)
             {
-                //Debug.WriteLine(functionErrorMessage + $"The file '{fileNameAndPath}' does not exists");
-                Log(LogSeverityType.Error, functionErrorMessage + $"The file '{fileNameAndPath}' does not exists");
+                Logger?.LogError("Failed to parse file '{fileNameAndPath}': the file does not exists", fileNameAndPath);
                 return false;
             }
 
             if (!fileInfo.Extension.EndsWith("igc"))
             {
-                //Debug.WriteLine(functionErrorMessage + $"The file extension '{fileInfo.Extension}' is not supported");
-                Log(LogSeverityType.Error, functionErrorMessage + $"The file extension '{fileInfo.Extension}' is not supported");
+                Logger?.LogError("Failed to parse file '{fileNameAndPath}': the file extension '{fileInfo.Extension}' is not supported", fileNameAndPath, fileInfo.Extension);
                 return false;
             }
 
             track = new Track();
             int pilotNumber = -1;
             string pilotIdentifier = "";
-            DateTime date = new DateTime();
+            DateTime date = new();
             int goalNortingDigits = -1;
             int goalEastingDigits = -1;
             bool declaredAltitudeIsInFeet = true;
-            List<string> lines = new List<string>();
-            using (StreamReader reader = new StreamReader($@"{fileNameAndPath}"))
+            List<string> lines = [];
+            using (StreamReader reader = new($@"{fileNameAndPath}"))
             {
                 while (!reader.EndOfStream)
                 {
@@ -61,14 +58,12 @@ namespace Coordinates.Parsers
             string[] identifierLine = lines.Where(x => x.StartsWith("AXXX")).ToArray();
             if (identifierLine.Length == 0)
             {
-                //Debug.WriteLine("Line with Pilot Identifier 'AXXX' is missing");
-                Log(LogSeverityType.Error, "Line with Pilot Identifier 'AXXX' is missing");
+                Logger?.LogError("Line with Pilot Identifier 'AXXX' is missing");
                 return false;
             }
             if (identifierLine.Length > 1)
             {
-                //Debug.WriteLine("More the one line with Pilot Identifier is found. First occurrence will be used.");
-                Log(LogSeverityType.Error, "More the one line with Pilot Identifier is found. First occurrence will be used.");
+                Logger?.LogError("More the one line with Pilot Identifier were found. First occurrence will be used");
             }
             pilotIdentifier = identifierLine[0].Replace("AXXX", "").Replace("Balloon Competition Logger", "");
 
@@ -78,25 +73,19 @@ namespace Coordinates.Parsers
                 if (headerLine.StartsWith("HFDTE"))
                 {
                     string line = headerLine.Replace("HFDTE", "");
-                    int day;
-                    if (!int.TryParse(line[0..2], out day))
+                    if (!int.TryParse(line[0..2], out int day))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse day portion of date '{line[0..2]}' in '{line}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse day portion of date '{line[0..2]}' in '{line}'");
+                        Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the day portion for date form '{portion}' in line '{line}'", fileNameAndPath, line[0..2], line);
                         return false;
                     }
-                    int month;
-                    if (!int.TryParse(line[2..4], out month))
+                    if (!int.TryParse(line[2..4], out int month))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse month portion of date '{line[2..4]}' in '{line}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse month portion of date '{line[2..4]}' in '{line}'");
+                        Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the month portion for date form '{portion}' in line '{line}'", fileNameAndPath, line[2..4], line);
                         return false;
                     }
-                    int year;
-                    if (!int.TryParse(line[4..6], out year))
+                    if (!int.TryParse(line[4..6], out int year))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse year portion of date '{line[4..6]}' in '{line}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse year portion of date '{line[4..6]}' in '{line}'");
+                        Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the year portion for date form '{portion}' in line '{line}'", fileNameAndPath, line[4..6], line);
                         return false;
                     }
                     year += 2000;
@@ -107,8 +96,7 @@ namespace Coordinates.Parsers
                     string pilotNumberText = headerLine.Replace("HFPID", "");
                     if (!int.TryParse(pilotNumberText, out pilotNumber))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse the pilot number '{pilotNumberText}' in '{headerLine}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse the pilot number '{pilotNumberText}' in '{headerLine}'");
+                        Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the pilot number '{pilotNumberText}' in line '{headerLine}'", fileNameAndPath, pilotNumberText, headerLine);
                         return false;
                     }
                 }
@@ -119,33 +107,19 @@ namespace Coordinates.Parsers
             {
                 if (configLine.StartsWith("LXXX declaration digits"))
                 {
-                    int digits;
-                    if (!int.TryParse(configLine[^1..^0], out digits))
+                    if (!int.TryParse(configLine[^1..^0], out int digits))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse number of declaration digits '{configLine[^1..^0]}' in '{configLine}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse number of declaration digits '{configLine[^1..^0]}' in '{configLine}'");
+
+                        Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the number of declaration digits '{declarationDigits}' in line '{configLine}'", fileNameAndPath, configLine[^1..^0], configLine);
                         return false;
                     }
-                    //TODO clarify with Mike
-                    //digits--;//subtract separator
+
                     goalNortingDigits = digits / 2;
                     goalEastingDigits = digits / 2;
                     if (digits % 2 == 1)//if digits are odd, expect that northing will declared with one digits more
                     {
                         goalNortingDigits++;
                     }
-
-                    ////TODO parse the different config value (should be value -1 but wait for conformation from Mike)
-                    //if (!int.TryParse(configLine[^3..^2], out goalNortingDigits))
-                    //{
-                    //    Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration northing digits '{configLine[^3..^2]}' in '{configLine}'");
-                    //    return false;
-                    //}
-                    //if (!int.TryParse(configLine[^1..^0], out goalEastingDigits))
-                    //{
-                    //    Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration easting digits '{configLine[^1..^0]}' in '{configLine}'");
-                    //    return false;
-                    //}
                 }
                 if (configLine.StartsWith("LXXX alt unit"))
                 {
@@ -158,22 +132,14 @@ namespace Coordinates.Parsers
                         declaredAltitudeIsInFeet = false;
                     }
                 }
-                if (configLine.StartsWith("LXXX loggerinterval"))
-                {
-                    string loggerInterval = configLine.Split("=").Last();
-                    if (!track.AdditionalPropertiesFromIGCFile.ContainsKey("LoggerInterval"))
-                        track.AdditionalPropertiesFromIGCFile.Add("LoggerInterval", loggerInterval.Replace("s",""));
-                }
             }
 
             string[] trackPointLines = lines.Where(x => x.StartsWith('B')).ToArray();
             foreach (string trackPointLine in trackPointLines)
             {
-                Coordinate coordinate;
-                if (!ParseTrackPoint(trackPointLine, date, out coordinate))
+                if (!ParseTrackPoint(trackPointLine, date, out Coordinate coordinate))
                 {
-                    //Debug.WriteLine(functionErrorMessage + "Failed to parse trackpoint");
-                    Log(LogSeverityType.Error, functionErrorMessage + "Failed to parse trackpoint");
+                    Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the trackpoint '{trackPointLine}'", fileNameAndPath, trackPointLine);
                     return false;
                 }
                 track.TrackPoints.Add(coordinate);
@@ -182,33 +148,31 @@ namespace Coordinates.Parsers
             string[] markerDropLines = lines.Where(x => x.StartsWith('E') && x.Contains("XX0")).ToArray();
             foreach (string markerDropLine in markerDropLines)
             {
-                MarkerDrop markerDrop;
-                if (!ParseMarkerDrop(markerDropLine, date, out markerDrop))
+                if (!ParseMarkerDrop(markerDropLine, date, out MarkerDrop markerDrop))
                 {
-                    //Debug.WriteLine(functionErrorMessage + "Failed to parse marker drop");
-                    Log(LogSeverityType.Error, functionErrorMessage + "Failed to parse marker drop");
+                    Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the marker drop '{markerDropLine}'", fileNameAndPath, markerDropLine);
                     return false;
                 }
                 track.MarkerDrops.Add(markerDrop);
             }
 
-            Coordinate referenceCoordinate = null;
-            if (track.MarkerDrops.Count > 0)
+            if (referenceCoordinate is null)
             {
-                MarkerDrop markerDrop = track.MarkerDrops.Find(x => x.MarkerNumber == 1);
-                if (markerDrop != null)
-                    referenceCoordinate = markerDrop.MarkerLocation;
+                if (track.MarkerDrops.Count > 0)
+                {
+                    MarkerDrop markerDrop = track.MarkerDrops.Find(x => x.MarkerNumber == 1);
+                    if (markerDrop != null)
+                        referenceCoordinate = markerDrop.MarkerLocation;
+                    else
+                    {
+                        referenceCoordinate = track.MarkerDrops[0].MarkerLocation;
+                        Logger?.LogWarning("No marker 1 dropped. Marker '{track.MarkerDrops[0].MarkerNumber}' will be used as goal declaration reference", track.MarkerDrops[0].MarkerNumber);
+                    }
+                }
                 else
                 {
-                    referenceCoordinate = track.MarkerDrops[0].MarkerLocation;
-                    //Debug.WriteLine($"No marker 1 dropped. Marker '{track.MarkerDrops[0].MarkerNumber}' will be used as goal declaration reference");
-                    Log(LogSeverityType.Warning, $"No marker 1 dropped. Marker '{track.MarkerDrops[0].MarkerNumber}' will be used as goal declaration reference");
+                    Logger?.LogWarning("No marker drops found. Position at declaration will be used as reference instead");
                 }
-            }
-            else
-            {
-                //Debug.WriteLine("No marker drops found. Position at declaration will be used as reference instead");
-                Log(LogSeverityType.Warning, "No marker drops found. Position at declaration will be used as reference instead");
             }
             string[] goalDeclarationLines = lines.Where(x => x.StartsWith('E') && x.Contains("XX1")).ToArray();
             foreach (string goalDeclarationLine in goalDeclarationLines)
@@ -224,8 +188,7 @@ namespace Coordinates.Parsers
                         {
                             if (!ParseTrackPoint(lines[index], date, out positionAtDeclaration))
                             {
-                                //Debug.WriteLine(functionErrorMessage + "Failed to parse track point as position of declaration");
-                                Log(LogSeverityType.Error, functionErrorMessage + "Failed to parse track point as position of declaration");
+                                Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the trackpoint as position of declaration '{lines[index]}'", fileNameAndPath, lines[index]);
                                 return false;
                             }
                             index = 0;
@@ -234,149 +197,19 @@ namespace Coordinates.Parsers
                     }
                     while (index > 0);
                 }
-                if (positionAtDeclaration == null)
+                if (positionAtDeclaration is null)
                 {
-                    //Debug.WriteLine(functionErrorMessage + "Failed to find a trackpoint as position of declaration");
-                    Log(LogSeverityType.Error, functionErrorMessage + "Failed to find a trackpoint as position of declaration");
+                    Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot find a trackpoint as position of declaration", fileNameAndPath);
                     return false;
                 }
-                Declaration declaration;
-                if (!ParseGoalDeclaration(goalDeclarationLine, date,
-                    declaredAltitudeIsInFeet,
-                    defaultGoalAltitude,
-                    goalNortingDigits,
-                    goalEastingDigits,
-                    referenceCoordinate,
-                    positionAtDeclaration,
-                    out declaration))
+                if (!ParseGoalDeclaration(goalDeclarationLine, date, declaredAltitudeIsInFeet, goalNortingDigits, goalEastingDigits, referenceCoordinate, positionAtDeclaration, out Declaration declaration))
                 {
-                    //Debug.WriteLine(functionErrorMessage + "Failed to parse goal declaration");
-                    Log(LogSeverityType.Error, functionErrorMessage + "Failed to parse goal declaration");
+                    Logger?.LogError("Failed to parse file '{fileNameAndPath}': cannot parse the goal declaration '{goalDeclarationLine}'", fileNameAndPath, goalDeclarationLine);
                     return false;
                 }
                 track.Declarations.Add(declaration);
             }
-            //using (StreamReader reader = new StreamReader(fileNameAndPath))
-            //{
-            //    while (!reader.EndOfStream)
-            //    {
-            //        string line = reader.ReadLine();
-            //        if (string.IsNullOrEmpty(line))
-            //            continue;
-            //        char lineIdenticator = line[0];
-            //        switch (lineIdenticator)
-            //        {
-            //            case 'A':
-            //                pilotIdentifier = line.Replace("AXXX", "").Replace("BalloonLive", "");
-            //                break;
-            //            case 'H':
-            //                if (line.StartsWith("HFDTE"))
-            //                {
-            //                    line = line.Replace("HFDTE", "");
-            //                    int day;
-            //                    if (!int.TryParse(line[0..2], out day))
-            //                    {
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse day portion of date '{line[0..2]}' in '{line}'");
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse day portion of date '{line[0..2]}' in '{line}'");
-            //                        return false;
-            //                    }
-            //                    int month;
-            //                    if (!int.TryParse(line[2..4], out month))
-            //                    {
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse month portion of date '{line[2..4]}' in '{line}'");
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse month portion of date '{line[2..4]}' in '{line}'");
-            //                        return false;
-            //                    }
-            //                    int year;
-            //                    if (!int.TryParse(line[4..6], out year))
-            //                    {
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse year portion of date '{line[4..6]}' in '{line}'");
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse year portion of date '{line[4..6]}' in '{line}'");
-            //                        return false;
-            //                    }
-            //                    year += 2000;
-            //                    date = new DateTime(year, month, day);
-            //                }
-            //                if (line.StartsWith("HFPID"))
-            //                {
-            //                    string pilotNumberText = line.Replace("HFPID", "");
-            //                    if (!int.TryParse(pilotNumberText, out pilotNumber))
-            //                    {
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse the pilot number '{pilotNumberText}' in '{line}'");
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse the pilot number '{pilotNumberText}' in '{line}'");
-            //                        return false;
-            //                    }
-            //                }
-            //                break;
-            //            case 'L':
-            //                if (line.StartsWith("LXXX declaration digits"))
-            //                {
-            //                    if (!int.TryParse(line[^3..^2], out goalNortingDigits))
-            //                    {
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration northing digits '{line[^3..^2]}' in '{line}'");
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration northing digits '{line[^3..^2]}' in '{line}'");
-            //                        return false;
-            //                    }
-            //                    if (!int.TryParse(line[^1..^0], out goalEastingDigits))
-            //                    {
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration easting digits '{line[^1..^0]}' in '{line}'");
-            //                        Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration easting digits '{line[^1..^0]}' in '{line}'");
-            //                        return false;
-            //                    }
-            //                }
-            //                if (line.StartsWith("LXXX alt unit"))
-            //                {
-            //                    if (line.Contains("feet"))
-            //                    {
-            //                        declaredAltitudeIsInFeet = true;
-            //                    }
-            //                    else
-            //                    {
-            //                        declaredAltitudeIsInFeet = false;
-            //                    }
-            //                }
-
-            //                break;
-            //            case 'B':
-            //                Coordinate coordinate;
-            //                if (!ParseTrackPoint(line, date, out coordinate))
-            //                {
-            //                    Debug.WriteLine(functionErrorMessage + "Failed to parse trackpoint");
-            //                    Debug.WriteLine(functionErrorMessage + "Failed to parse trackpoint");
-            //                    return false;
-            //                }
-            //                track.TrackPoints.Add(coordinate);
-            //                break;
-            //            case 'E':
-            //                if (line.Contains("XL1"))
-            //                {
-            //                    DeclaredGoal declaredGoal;
-            //                    if (!ParseGoalDeclaration(line, date,declaredAltitudeIsInFeet, goalNortingDigits, goalEastingDigits, out declaredGoal))
-            //                    {
-            //                        Debug.WriteLine(functionErrorMessage + "Failed to parse goal declaration");
-            //                        Debug.WriteLine(functionErrorMessage + "Failed to parse goal declaration");
-            //                        return false;
-            //                    }
-            //                    track.DeclaredGoals.Add(declaredGoal);
-            //                }
-            //                if (line.Contains("XX0"))
-            //                {
-            //                    MarkerDrop markerDrop;
-            //                    if (!ParseMarkerDrop(line, date, out markerDrop))
-            //                    {
-            //                        Debug.WriteLine(functionErrorMessage + "Failed to parse marker drop");
-            //                        Debug.WriteLine(functionErrorMessage + "Failed to parse marker drop");
-            //                        return false;
-            //                    }
-            //                    track.MarkerDrops.Add(markerDrop);
-            //                }
-            //                break;
-            //            default:
-            //                break;
-            //        }
-            //    }
-            //}
-            Pilot pilot = new Pilot(pilotNumber, pilotIdentifier);
+            Pilot pilot = new(pilotNumber, pilotIdentifier);
             track.Pilot = pilot;
             return true;
         }
@@ -397,46 +230,35 @@ namespace Coordinates.Parsers
         /// <returns>true:success; false:error</returns>
         private static bool ParseTrackPoint(string line, DateTime date, out Coordinate coordinate)
         {
-            string functionErrorMessage = $"Failed to parse track point:";
             coordinate = null;
 
-            DateTime timeStamp;
-            if (!ParseTimeStamp(line, date, out timeStamp))
+            if (!ParseTimeStamp(line, date, out DateTime timeStamp))
             {
-                //Debug.WriteLine(functionErrorMessage);
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse track point: cannot parse the time stamp in line '{line}'", line);
                 return false;
             }
 
-            double latitude;
-            if (!ParseLatitude(line[7..15], out latitude))
+            if (!ParseLatitude(line[7..15], out double latitude))
             {
-                //Debug.WriteLine(functionErrorMessage);
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse track point: cannot parse the latitude portion '{portion}' in line '{line}'", line[7..15], line);
                 return false;
             }
-            double longitude;
-            if (!ParseLongitude(line[15..24], out longitude))
+            if (!ParseLongitude(line[15..24], out double longitude))
             {
-                //Debug.WriteLine(functionErrorMessage);
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse track point: cannot parse the longitude portion '{portion}' in line '{line}'", line[15..24], line);
                 return false;
             }
 
 
-            double altitudeBarometric;
-            if (!double.TryParse(line[25..30], out altitudeBarometric))
+            if (!double.TryParse(line[25..30], out double altitudeBarometric))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude '{line[25..30]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse barometric altitude '{line[25..30]}' in '{line}'");
+                Logger?.LogError("Failed to parse track point: cannot parse the barometric altitude portion '{portion}' in line '{line}'", line[25..30], line);
                 return false;
             }
 
-            double altitudeGPS;
-            if (!double.TryParse(line[30..35], out altitudeGPS))
+            if (!double.TryParse(line[30..35], out double altitudeGPS))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude '{line[30..35]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse GPS altitude '{line[30..35]}' in '{line}'");
+                Logger?.LogError("Failed to parse track point: cannot parse the GPS altitude portion '{portion}' in line '{line}'", line[30..35], line);
                 return false;
             }
             coordinate = new Coordinate(latitude, longitude, altitudeGPS, altitudeBarometric, timeStamp);
@@ -458,57 +280,46 @@ namespace Coordinates.Parsers
         /// <param name="referenceCoordinate">a reference coordinate to fill up the missing info from utm goal declaration. If the reference is null, the position of declaration will be used instead</param>
         /// <param name="declaration">output parameter. the declared goal</param>
         /// <returns>true:success; false:error</returns>
-        private static bool ParseGoalDeclaration(string line, DateTime date, bool declaredAltitudeIsInFeet, double defaultGoalAltiude, int northingDigits, int eastingDigits, Coordinate referenceCoordinate, Coordinate positionAtDeclaration, out Declaration declaration)
+        private static bool ParseGoalDeclaration(string line, DateTime date, bool declaredAltitudeIsInFeet, int northingDigits, int eastingDigits, Coordinate referenceCoordinate, Coordinate positionAtDeclaration, out Declaration declaration)
         {
-            string functionErrorMessage = $"Failed to parse goal declaration:";
             declaration = null;
 
-            DateTime timeStamp;
-            if (!ParseTimeStamp(line, date, out timeStamp))
+            if (!ParseTimeStamp(line, date, out DateTime timeStamp))
             {
-                //Debug.WriteLine(functionErrorMessage);
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse goal declaration: cannot parse the time stamp in line '{line}'", line);
                 return false;
             }
-            int goalNumber;
-            if (!int.TryParse(line[10..12], out goalNumber))
+            if (!int.TryParse(line[10..12], out int goalNumber))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal number '{line[10..12]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal number '{line[10..12]}' in '{line}'");
+                Logger?.LogError("Failed to parse goal declaration: cannot parse the goal number portion '{portion}' in line '{line}'", line[10..12], line);
                 return false;
             }
-            int eastingUTM;
             //int eastingStartCharacter = northingStartCharacter + northingDigits + 1;
             int eastingStartCharacter = 12;
-            if (!int.TryParse(line[eastingStartCharacter..(eastingStartCharacter + eastingDigits)], out eastingUTM))
+            if (!int.TryParse(line[eastingStartCharacter..(eastingStartCharacter + eastingDigits)], out int originalEastingDeclarationUTM))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration easting portion '{line[eastingStartCharacter..(eastingStartCharacter + eastingDigits)]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal declaration easting portion '{line[eastingStartCharacter..(eastingStartCharacter + eastingDigits)]}' in '{line}'");
+                Logger?.LogError("Failed to parse goal declaration: cannot parse the easting portion '{portion}' in line '{line}'", line[eastingStartCharacter..(eastingStartCharacter + eastingDigits)], line);
                 return false;
             }
-            int northingUTM;
             //int northingStartCharacter = 12;
             int northingStartCharacter = eastingStartCharacter + eastingDigits + 1;
-            if (!int.TryParse(line[northingStartCharacter..(northingStartCharacter + northingDigits)], out northingUTM))
+            if (!int.TryParse(line[northingStartCharacter..(northingStartCharacter + northingDigits)], out int originalNorthingDeclarationUTM))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration northing portion '{line[northingStartCharacter..(northingStartCharacter + northingDigits)]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal declaration northing portion '{line[northingStartCharacter..(northingStartCharacter + northingDigits)]}' in '{line}'");
+                Logger?.LogError("Failed to parse goal declaration: cannot parse the northing portion '{portion}' in line '{line}'", line[northingStartCharacter..(northingStartCharacter + northingDigits)], line);
                 return false;
             }
             string[] parts = line.Split(',');
-            int declaredAltitude;
             double declaredAltitudeInMeter;
-            //TODO clarify with Mike
+            bool hasPilotDeclaredGoalAltitude = true;
             if (parts.Length > 1)
             {
                 if (!string.IsNullOrWhiteSpace(parts[1]))
                 {
                     parts[1] = parts[1].Replace("ft", "").Replace("m", "");
 
-                    if (!int.TryParse(parts[1], out declaredAltitude))
+                    if (!int.TryParse(parts[1], out int declaredAltitude))
                     {
-                        //Debug.WriteLine(functionErrorMessage + $"Failed to parse goal declaration altitude portion '{parts[1]}' in '{line}'");
-                        Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse goal declaration altitude portion '{parts[1]}' in '{line}'");
+                        Logger?.LogError("Failed to parse goal declaration: cannot parse the altitude portion '{portion}' in line '{line}'", parts[1], line);
                         return false;
                     }
                     if (declaredAltitudeIsInFeet)
@@ -518,17 +329,18 @@ namespace Coordinates.Parsers
                 }
                 else
                 {
-
-                    declaredAltitudeInMeter = defaultGoalAltiude;
-                    Log(LogSeverityType.Warning, $"No altitude declared for Goal No. '{goalNumber}'. Default gaol altitude of '{defaultGoalAltiude}' will be assumed");
+                    hasPilotDeclaredGoalAltitude = false;
+                    declaredAltitudeInMeter = 0.0;
+                    Logger?.LogWarning("No altitude declared for Goal No. '{goalNumber}'. Altitude of 0 will be assumed", goalNumber);
                 }
             }
             else
             {
-
-                declaredAltitudeInMeter = defaultGoalAltiude;
-                Log(LogSeverityType.Warning, $"No altitude declared for Goal No. '{goalNumber}'. Default gaol altitude of '{defaultGoalAltiude}' will be assumed");
+                hasPilotDeclaredGoalAltitude = false;
+                declaredAltitudeInMeter = 0.0;
+                Logger?.LogWarning("No altitude declared for Goal No. '{goalNumber}'. Altitude of 0 will be assumed", goalNumber);
             }
+
 
             CoordinateSharp.Coordinate coordinateSharp;
             if (referenceCoordinate == null)
@@ -537,6 +349,8 @@ namespace Coordinates.Parsers
                 coordinateSharp = new CoordinateSharp.Coordinate(referenceCoordinate.Latitude, referenceCoordinate.Longitude);
 
             string utmGridZone = coordinateSharp.UTM.LatZone + coordinateSharp.UTM.LongZone;
+            int northingUTM = originalNorthingDeclarationUTM;
+            int eastingUTM = originalEastingDeclarationUTM;
             if (northingDigits < 6)
             {
                 northingUTM *= 10;
@@ -553,14 +367,13 @@ namespace Coordinates.Parsers
                 eastingUTM += (int)(Math.Floor(coordinateSharp.UTM.Easting / Math.Pow(10, eastingDigits + 1)) * Math.Pow(10, eastingDigits + 1));
             }
 
-            CoordinateSharp.UniversalTransverseMercator utm = new CoordinateSharp.UniversalTransverseMercator(utmGridZone, eastingUTM, northingUTM);
+            CoordinateSharp.UniversalTransverseMercator utm = new(utmGridZone, eastingUTM, northingUTM);
 
             CoordinateSharp.Coordinate coordinate = CoordinateSharp.UniversalTransverseMercator.ConvertUTMtoLatLong(utm);
 
-            Coordinate declaredGoal = new Coordinate(coordinate.Latitude.DecimalDegree, coordinate.Longitude.DecimalDegree, declaredAltitudeInMeter, declaredAltitudeInMeter, timeStamp);
-            //Coordinate positionAtDeclaration = new Coordinate(positionAtDeclaration.Latitude, positionAtDeclaration.Longitude, positionAtDeclaration.AltitudeGPS, positionAtDeclaration.AltitudeBarometric, positionAtDeclaration.TimeStamp);
+            Coordinate declaredGoal = new(coordinate.Latitude.DecimalDegree, coordinate.Longitude.DecimalDegree, declaredAltitudeInMeter, declaredAltitudeInMeter, timeStamp);
 
-            declaration = new Declaration(goalNumber, declaredGoal, positionAtDeclaration);
+            declaration = new Declaration(goalNumber, declaredGoal, positionAtDeclaration, hasPilotDeclaredGoalAltitude, originalEastingDeclarationUTM, originalNorthingDeclarationUTM);
 
             return true;
         }
@@ -577,55 +390,42 @@ namespace Coordinates.Parsers
         /// <returns>true:success; false: error</returns>
         private static bool ParseMarkerDrop(string line, DateTime date, out MarkerDrop markerDrop)
         {
-            string functionErrorMessage = $"Failed to parse marker drop:";
             markerDrop = null;
 
-            DateTime timeStamp;
-            if (!ParseTimeStamp(line, date, out timeStamp))
+            if (!ParseTimeStamp(line, date, out DateTime timeStamp))
             {
-                //Debug.WriteLine(functionErrorMessage);
-                Log(LogSeverityType.Error, functionErrorMessage);
+                Logger?.LogError("Failed to parse marker drop: cannot parse the time stamp in line '{line}'", line);
                 return false;
             }
-            int markerNumber;
-            if (!int.TryParse(line[10..12], out markerNumber))
+            if (!int.TryParse(line[10..12], out int markerNumber))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse marker number '{line[10..12]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse marker number '{line[10..12]}' in '{line}'");
+                Logger?.LogError("Failed to parse marker drop: cannot parse the marker number portion '{portion}' in line '{line}'", line[10..12], line);
                 return false;
             }
 
-            double latitude;
-            if (!ParseLatitude(line[12..20], out latitude))
+            if (!ParseLatitude(line[12..20], out double latitude))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse marker drop latitude '{line[12..20]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse marker drop latitude '{line[12..20]}' in '{line}'");
+                Logger?.LogError("Failed to parse marker drop: cannot parse the latitude portion '{portion}' in line '{line}'", line[12..20], line);
                 return false;
             }
-            double longitude;
-            if (!ParseLongitude(line[20..29], out longitude))
+            if (!ParseLongitude(line[20..29], out double longitude))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse marker drop longitude '{line[20..29]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse marker drop longitude '{line[20..29]}' in '{line}'");
+                Logger?.LogError("Failed to parse marker drop: cannot parse the longitude portion '{portion}' in line '{line}'", line[20..29], line);
                 return false;
             }
 
-            double altitudeBarometric;
-            if (!double.TryParse(line[30..35], out altitudeBarometric))
+            if (!double.TryParse(line[30..35], out double altitudeBarometric))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude '{line[30..35]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse barometric altitude '{line[30..35]}' in '{line}'");
+                Logger?.LogError("Failed to parse marker drop: cannot parse the barometric altitude portion '{portion}' in line '{line}'", line[30..35], line);
                 return false;
             }
 
-            double altitudeGPS;
-            if (!double.TryParse(line[35..40], out altitudeGPS))
+            if (!double.TryParse(line[35..40], out double altitudeGPS))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse barometric altitude '{line[35..40]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse barometric altitude '{line[35..40]}' in '{line}'");
+                Logger?.LogError("Failed to parse marker drop: cannot parse the GPS altitude portion '{portion}' in line '{line}'", line[35..40], line);
                 return false;
             }
-            Coordinate coordinate = new Coordinate(latitude, longitude, altitudeGPS, altitudeBarometric, timeStamp);
+            Coordinate coordinate = new(latitude, longitude, altitudeGPS, altitudeBarometric, timeStamp);
             markerDrop = new MarkerDrop(markerNumber, coordinate);
             return true;
         }
@@ -646,23 +446,17 @@ namespace Coordinates.Parsers
                 factor = -1.0;
             else
             {
-                factor = double.NaN;
-                //Debug.WriteLine($"Failed to parse latitude text. Unexpected suffix '{latitudeText[^1]}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text. Unexpected suffix '{latitudeText[^1]}'");
+                Logger?.LogError("Failed to parse latitude text. Unexpected suffix '{latitudeText[^1]}'", latitudeText[^1]);
                 return false;
             }
-            double fullAngle = double.NaN;
-            if (!double.TryParse(latitudeText[0..2], out fullAngle))
+            if (!double.TryParse(latitudeText[0..2], out double fullAngle))
             {
-                //Debug.WriteLine($"Failed to parse latitude text '{latitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text '{latitudeText}'");
+                Logger?.LogError("Failed to parse latitude full angle portion '{portion}' from '{latitudeText}'", latitudeText[0..2], latitudeText);
                 return false;
             }
-            double decimalAngle = double.NaN;
-            if (!double.TryParse(latitudeText[2..7], out decimalAngle))
+            if (!double.TryParse(latitudeText[2..7], out double decimalAngle))
             {
-                //Debug.WriteLine($"Failed to parse latitude text '{latitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse latitude text '{latitudeText}'");
+                Logger?.LogError("Failed to parse latitude decimal angle portion '{portion}' from '{latitudeText}'", latitudeText[2..7], latitudeText);
                 return false;
             }
             decimalAngle /= 60000.0;//divided by 1000 to get decimal value, divided by 60 to get from angle minutes to decimal angles
@@ -687,23 +481,17 @@ namespace Coordinates.Parsers
                 factor = -1.0;
             else
             {
-                factor = double.NaN;
-                //Debug.WriteLine($"Failed to parse longitude text. Unexpected suffix '{longitudeText[^1]}'");
-                Log(LogSeverityType.Error, $"Failed to parse longitude text. Unexpected suffix '{longitudeText[^1]}'");
+                Logger?.LogError("Failed to parse longitude text. Unexpected suffix '{longitudeText[^1]}'", longitudeText[^1]);
                 return false;
             }
-            double fullAngle = double.NaN;
-            if (!double.TryParse(longitudeText[0..3], out fullAngle))
+            if (!double.TryParse(longitudeText[0..3], out double fullAngle))
             {
-                //Debug.WriteLine($"Failed to parse longitude text '{longitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse longitude text '{longitudeText}'");
+                Logger?.LogError("Failed to parse longitude full angle portion '{portion}' from '{longitudeText}'", longitudeText[0..3], longitudeText);
                 return false;
             }
-            double decimalAngle = double.NaN;
-            if (!double.TryParse(longitudeText[3..8], out decimalAngle))
+            if (!double.TryParse(longitudeText[3..8], out double decimalAngle))
             {
-                //Debug.WriteLine($"Failed to parse longitude text '{longitudeText}'");
-                Log(LogSeverityType.Error, $"Failed to parse longitude text '{longitudeText}'");
+                Logger?.LogError("Failed to parse longitude decimal angle portion '{portion}' from '{longitudeText}'", longitudeText[3..8], longitudeText);
                 return false;
             }
             decimalAngle /= 60000.0;//divided by 1000 to get decimal value, divided by 60 to get from angle minutes to decimal angles
@@ -721,37 +509,25 @@ namespace Coordinates.Parsers
         /// <returns>true:success: false:error</returns>
         private static bool ParseTimeStamp(string line, DateTime date, out DateTime timeStamp)
         {
-            string functionErrorMessage = "Failed to parse time:";
             string time = line[1..7];
-            int hours;
             timeStamp = date;
-            if (!int.TryParse(time[0..2], out hours))
+            if (!int.TryParse(time[0..2], out int hours))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse hour portion '{time[0..2]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse hour portion '{time[0..2]}' in '{line}'");
+                Logger?.LogError("Failed to parse hour portion '{portion}' in line '{line}'", time[0..2], line);
                 return false;
             }
-            int minutes;
-            if (!int.TryParse(time[2..4], out minutes))
+            if (!int.TryParse(time[2..4], out int minutes))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse minute portion '{time[2..4]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse minute portion '{time[2..4]}' in '{line}'");
+                Logger?.LogError("Failed to parse minute portion '{portion}' in line '{line}'", time[2..4], line);
                 return false;
             }
-            int seconds;
-            if (!int.TryParse(time[4..6], out seconds))
+            if (!int.TryParse(time[4..6], out int seconds))
             {
-                //Debug.WriteLine(functionErrorMessage + $"Failed to parse second portion '{time[4..6]}' in '{line}'");
-                Log(LogSeverityType.Error, functionErrorMessage + $"Failed to parse second portion '{time[4..6]}' in '{line}'");
+                Logger?.LogError("Failed to parse second portion '{portion}' in line '{line}'", time[4..6], line);
                 return false;
             }
             timeStamp = date.AddHours(hours).AddMinutes(minutes).AddSeconds(seconds);
             return true;
-        }
-
-        private static void Log(LogSeverityType logSeverity, string text)
-        {
-            Logger.Log((object)"FAI Logger Parser", logSeverity, text);
         }
         #endregion
     }
