@@ -1,13 +1,8 @@
 ﻿using Coordinates;
-using LoggerComponent;
+using LoggingConnector;
+using Microsoft.Extensions.Logging;
 using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
 using System.IO;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -16,9 +11,10 @@ namespace TrackReportGenerator
     public partial class TrackReportGeneratorForm : Form
     {
         #region Properties
-        private bool UseGPSAltitude=true;
-        private double MaxAllowedAltitude=CoordinateHelpers.ConvertToMeter(10000);
-        private bool SkipCoordinatesWithoutLocation=true;
+        private readonly ILogger<TrackReportGeneratorForm> Logger=LogConnector.LoggerFactory.CreateLogger<TrackReportGeneratorForm>();
+        private bool UseGPSAltitude = true;
+        private double MaxAllowedAltitude = CoordinateHelpers.ConvertToMeter(10000);
+        private bool SkipCoordinatesWithoutLocation = true;
         #endregion Properties
 
         #region Constructor
@@ -26,7 +22,7 @@ namespace TrackReportGenerator
         {
             InitializeComponent();
             Text += typeof(TrackReportGeneratorForm).Assembly.GetName().Version;
-            logListView1.StartLogging(@".\Logfile.txt");
+            //logListView1.StartLogging(@".\Logfile.txt");
         }
 
         #endregion Constructor
@@ -36,7 +32,7 @@ namespace TrackReportGenerator
         private async void btSelectFiles_Click(object sender, EventArgs e)
         {
             progressBar1.Value = 0;
-            OpenFileDialog openFileDialog = new OpenFileDialog();
+            OpenFileDialog openFileDialog = new();
             if (!string.IsNullOrWhiteSpace(Properties.Settings.Default.InitalDirectory))
                 openFileDialog.InitialDirectory = Properties.Settings.Default.InitalDirectory;
             openFileDialog.Title = "Select .igc files";
@@ -88,13 +84,12 @@ namespace TrackReportGenerator
                 }
                 if (rbBallonLiveParser.Checked)
                 {
-                    string changeOfPositionSource;
-                    if (track.AdditionalPropertiesFromIGCFile.TryGetValue("Change of position source", out changeOfPositionSource))
+                    if (track.AdditionalPropertiesFromIGCFile.TryGetValue("Change of position source", out string changeOfPositionSource))
                     {
-                        if (changeOfPositionSource.ToLower() == "yes")
+                        if (changeOfPositionSource.Equals("yes", StringComparison.CurrentCultureIgnoreCase))
                         {
-                            Logger.Log(LogSeverityType.Warning, "Caution: change of position source has been detected, refer the log for more details");
-                         //   MessageBox.Show("Caution: change of position source has been detected, refer the log for more details");
+                            Logger?.LogWarning("Caution: change of position source has been detected, refer the log for more details");
+                            //   MessageBox.Show("Caution: change of position source has been detected, refer the log for more details");
                         }
                     }
                 }
@@ -106,7 +101,7 @@ namespace TrackReportGenerator
                 }
                 else
                 {
-                    Logger.Log(LogSeverityType.Info, $"File '{Path.GetFileName(igcFile)}' skipped");
+                    Logger?.LogInformation("File '{fileName}' skipped", Path.GetFileName(igcFile));
                 }
 
                 return true;
@@ -118,7 +113,7 @@ namespace TrackReportGenerator
             UseGPSAltitude = rbGPSAltitude.Checked;
         }
 
-        
+
 
         private void rbMeter_CheckedChanged(object sender, EventArgs e)
         {
@@ -128,7 +123,7 @@ namespace TrackReportGenerator
             }
             else
             {
-                tbMaxAltitude.Text = $"{Math.Round(CoordinateHelpers.ConvertToFeet(MaxAllowedAltitude),0,MidpointRounding.AwayFromZero)}";
+                tbMaxAltitude.Text = $"{Math.Round(CoordinateHelpers.ConvertToFeet(MaxAllowedAltitude), 0, MidpointRounding.AwayFromZero)}";
             }
         }
 
@@ -160,10 +155,9 @@ namespace TrackReportGenerator
 
         private void tbMaxAltitude_Leave(object sender, EventArgs e)
         {
-            double tempMaxAltitude;
-            if (!double.TryParse(tbMaxAltitude.Text, out tempMaxAltitude))
+            if (!double.TryParse(tbMaxAltitude.Text, out double tempMaxAltitude))
             {
-                Logger.Log(LogSeverityType.Error, $"Failed to parse '{tbMaxAltitude.Text}' as double. Please enter a number");
+                Logger?.LogError("Failed to parse '{maxAltitude}' as double. Please enter a number", tbMaxAltitude.Text);
                 return;
             }
             if (rbMeter.Checked)
