@@ -1,22 +1,17 @@
 ﻿using Coordinates;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace Shapes
 {
-    public class Sphere:Shapes3D
+    public class Sphere : Shapes3D
     {
         public Coordinate CenterPoint
         {
-            get;private set;
+            get; private set;
         }
 
         public double Radius
         {
-            get;private set;
+            get; private set;
         }
 
         public Sphere(Coordinate centerPoint, double radius)
