@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Coordinates
+﻿namespace Coordinates
 {
     public class Declaration
 
@@ -11,18 +7,37 @@ namespace Coordinates
         /// <summary>
         /// The number of the goal
         /// </summary>
-        public int GoalNumber { get; private set; }
+        public int GoalNumber
+        {
+            get; private set;
+        }
 
         /// <summary>
         /// The target or goal which has been declared
         /// </summary>
-        public Coordinate DeclaredGoal { get; private set; }
+        public Coordinate DeclaredGoal
+        {
+            get; private set;
+        }
 
+
+        public int OrignalEastingDeclarationUTM
+        {
+            get; private set;
+        }
+
+        public int OrignalNorhtingDeclarationUTM
+        {
+            get; private set;
+        }
 
         /// <summary>
         /// The position at which the goal has been declared
         /// </summary>
-        public Coordinate PositionAtDeclaration { get; private set; }
+        public Coordinate PositionAtDeclaration
+        {
+            get; private set;
+        }
 
         /// <summary>
         /// Creates a gaol
@@ -30,13 +45,14 @@ namespace Coordinates
         /// <param name="goalNumber">The number of the goal</param>
         /// <param name="declaredGoal">The target or goal which has been declared</param>
         /// <param name="positionAtDeclaration">The position at which the goal has been declared</param>
-        public Declaration(int goalNumber,Coordinate declaredGoal, Coordinate positionAtDeclaration)
+        public Declaration(int goalNumber, Coordinate declaredGoal, Coordinate positionAtDeclaration, bool hasPilotDelaredGoalAltitude, int orignalEastingDeclarationUTM, int orignalNorhtingDeclarationUTM)
         {
             GoalNumber = goalNumber;
             DeclaredGoal = declaredGoal;
             PositionAtDeclaration = positionAtDeclaration;
+            HasPilotDelaredGoalAltitude = hasPilotDelaredGoalAltitude;
+            OrignalEastingDeclarationUTM = orignalEastingDeclarationUTM;
+            OrignalNorhtingDeclarationUTM = orignalNorhtingDeclarationUTM;
         }
-
-        
     }
 }
