@@ -1,8 +1,10 @@
+using LoggingConnector;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Extensions.Logging;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows.Forms;
+using UILoggingProvider;
 
 namespace TrackReportGenerator
 {
@@ -17,6 +19,21 @@ namespace TrackReportGenerator
             Application.SetHighDpiMode(HighDpiMode.SystemAware);
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
+            var uiLoggerProvider = UILoggerProvider.Instance;
+            var builder = Host.CreateDefaultBuilder()
+                .ConfigureLogging(logging =>
+                    {
+                        logging.ClearProviders();
+                        logging.AddProvider(uiLoggerProvider);
+                    })
+                .ConfigureServices((_, services) =>
+                    {
+                        services.AddLogging();
+                    });
+
+            var host = builder.Build();
+            var loggerFactory = host.Services.GetRequiredService<ILoggerFactory>();
+            LogConnector.LoggerFactory = loggerFactory;
             Application.Run(new TrackReportGeneratorForm());
         }
     }
